% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcDefaultDrivers.R
\name{calcDefaultDrivers}
\alias{calcDefaultDrivers}
\title{Collect Default Model Drivers}
\usage{
calcDefaultDrivers(drivers = c("Population", "UrbanPop", "GDP", "GDPpc"))
}
\arguments{
\item{drivers}{Vector of strings.}
}
\value{
A magpie object with sets "region", "period" and "scen".
}
\description{
Collect Default Model Drivers
}
\examples{
\dontrun{
library(mrdrivers)
calcOutput("DefaultDrivers")}

}
\seealso{
\link[madrat:calcOutput]{madrat::calcOutput}
}
