% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcDriver.R
\name{calcDriver}
\alias{calcDriver}
\title{Vectorised scenario construction}
\usage{
calcDriver(
  driver,
  scenario,
  popAsWeight = FALSE,
  naming = "indicator_scenario",
  extension2150 = "bezier",
  ...
)
}
\arguments{
\item{driver}{A string designating the driver.}

\item{scenario}{A string designating the scenario.}

\item{popAsWeight}{If TRUE, then population data of the same scenario is used as weight.}

\item{naming}{A string giving the naming scheme of the data dimension. Can be either:
\itemize{
\item "indicator_scenario" (default): Returns names of the type "gdp_SSP2", or "pop_SSP2".
\item "indicator.scenario": Returns names of the type "gdp.SSP2", or "pop.SSP2".
\item "scenario": Returns names of the type "SSP2".
}
Set naming to "scenario" when you want to operate on SSP2 gdp and population data for instance, and not have to
worry about the conflicting names.}

\item{extension2150}{A string specifying if/how the scenarios should be extended until 2150. Can be either:
\itemize{
\item "bezier" (default): A bezier curve extension that leads to a smooth flattening of the scenario: the
slope in the last year of the scenario is halved by 2150. Currently only works for scenarios with 2100 as
their last year.
\item "constant": The last value of the scenarios is taken as constant until 2150.
\item "none": No extension.
}}

\item{...}{
  Arguments passed on to \code{\link[=calcScenarioConstructor]{calcScenarioConstructor}}
  \describe{
    \item{\code{harmonization}}{A string designating the harmonization function.}
    \item{\code{pastData}}{A string passed to the calc'Driver'Past function, e.g. \code{\link[=calcGDPPast]{calcGDPPast()}} or \code{\link[=calcPopulationPast]{calcPopulationPast()}}.}
    \item{\code{futureData}}{A string passed to the calc'Driver'Future function, e.g. \code{\link[=calcGDPFuture]{calcGDPFuture()}} or
\code{\link[=calcPopulationFuture]{calcPopulationFuture()}}.}
  }}
}
\value{
magpie object with the requested output data either on country or on
regional level depending on the choice of argument "aggregate" or a list of information
if supplementary is set to TRUE.
}
\description{
Vectorised scenario construction
}
\section{Vectorization of arguments}{
Vectors are accepted for -Calib, -Past and -Future arguments.
If given a vector, different combinations are created and returned all at once. If more than one
argument is vectorised, the arguments have to have the same length. Which time series are created can be
illustrated with the following example. Let's say the harmonization function and past data source are vectors of
length 3. Then there will be in total 3 time series that are produced: the first time series is the result of
combining the first harmonization function with the first past data source, the second time series the result of
combining the second harmonization function with the second past data source, and the third time series the result
of  using the respective third entry. The future data source used in each case is the same, since in this example
only one future data source is provided.
}

\keyword{internal}
