% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcGDP.R
\name{calcGDP}
\alias{calcGDP}
\alias{calcGDPpc}
\title{Get GDP and GDP per capita scenarios}
\usage{
calcGDP(
  scenario,
  unit = toolGetUnitDollar(inPPP = TRUE),
  average2020 = TRUE,
  ...
)

calcGDPpc(
  scenario,
  unit = toolGetUnitDollar(inPPP = TRUE),
  average2020 = TRUE,
  ...
)
}
\arguments{
\item{scenario}{A string (or vector of strings) designating the scenario(s) to be returned. Use
\code{\link[=toolGetScenarioDefinition]{toolGetScenarioDefinition()}} to learn what scenarios are available.}

\item{unit}{A string specifying the unit of GDP. Can be either:
\itemize{
\item "constant 2017 Int$PPP" (default):
Scenarios are constructed in constant 2017 Int$PPP.
\item "constant 2017 US$MER": Scenarios are constructed in constant 2017 Int$PPP and then
converted with \code{\link[GDPuc:toolConvertGDP]{GDPuc::toolConvertGDP()}}.
}
In all cases, GDP is returned in millions.}

\item{average2020}{If TRUE (default), then the 2020 value is replaced by the 2018-2022 average. To be consistent,
the yearly resolution is decreased to 5 year intervals.}

\item{...}{
  Arguments passed on to \code{\link{calcDriver}}
  \describe{
    \item{\code{extension2150}}{A string specifying if/how the scenarios should be extended until 2150. Can be either:
\itemize{
\item "bezier" (default): A bezier curve extension that leads to a smooth flattening of the scenario: the
slope in the last year of the scenario is halved by 2150. Currently only works for scenarios with 2100 as
their last year.
\item "constant": The last value of the scenarios is taken as constant until 2150.
\item "none": No extension.
}}
  }}
}
\value{
magpie object with the requested output data either on country or on
regional level depending on the choice of argument "aggregate" or a list of information
if supplementary is set to TRUE.
}
\description{
Like all scenarios in mrdrivers, the GDP and GDP per capita scenarios are the result of a harmonization exercise
between past data and future projections. Together with the corresponding population scenarios
(see \code{\link[=calcPopulation]{calcPopulation()}}) they comprise a consistent set of scenarios.

The \code{scenario} argument is used to designate the scenario(s) to be returned. Currently available GDP scenarios are:
SSPs, SSP1, SSP2, SSP3, SSP4, SSP5, SDPs, ISIMIP, SSP2IndiaDEAs, SSP2IndiaMedium, SSP2IndiaHigh. See the vignette: \code{vignette("scenarios")} and/or
\code{\link[=toolGetScenarioDefinition]{toolGetScenarioDefinition()}} for more information, scenario options, definitions and references.
}
\examples{
\dontrun{
# Return all SSP scenarios
calcOutput("GDP", scenario = "SSPs")

# Return only the SSP2 GDP scenario
calcOutput("GDP", scenario = "SSP2")

# Return the SSP and SDP GDPpc scenarios
calcOutput("GDP", scenario = c("SSPs", "SDPs"))
}

\dontrun{
calcOutput("GDPpc")
}

}
\seealso{
\itemize{
\item \code{\link[=toolGetScenarioDefinition]{toolGetScenarioDefinition()}} for scenario options and definitions.
\item \code{\link[madrat:calcOutput]{madrat::calcOutput()}} for how to return supplementary information and other control options.
\item \code{\link[=calcDriver]{calcDriver()}}, \code{\link[=calcScenarioConstructor]{calcScenarioConstructor()}} and \code{\link[=calcHarmonizedData]{calcHarmonizedData()}} for how to create new scenarios
(for developers).
}
}
