% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcGDPFuture.R
\name{calcGDPFuture}
\alias{calcGDPFuture}
\title{calcGDPFuture}
\usage{
calcGDPFuture(
  GDPFuture = "SSPs",
  unit = "constant 2005 Int$PPP",
  useMIData = TRUE,
  extension2150 = "bezier"
)
}
\arguments{
\item{GDPFuture}{GDP future data source}

\item{unit}{string}

\item{useMIData}{logical}

\item{extension2150}{string, either "bezier", "constant" or "none"}
}
\value{
A magpie object.
}
\description{
Calculates a time series of GDP in Purchase Power Parity (PPP) of million
International Dollars of the year 2005. Different sources are available:
\itemize{ \item \code{SSP}: Lavinia? \item \code{OECD}: Lavinia? }
}
