% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcGDPFuture.R, R/calcGDPPast.R,
%   R/calcGDPpcFuture.R, R/calcGDPpcPast.R
\name{calcGDPFuture}
\alias{calcGDPFuture}
\alias{calcGDPPast}
\alias{calcGDPpcFuture}
\alias{calcGDPpcPast}
\title{Get GDP and GDPpc scenario building blocks}
\usage{
calcGDPFuture(GDPFuture = "SSPs-MI", unit = "constant 2017 Int$PPP")

calcGDPPast(GDPPast = "WDI-MI", unit = "constant 2017 Int$PPP")

calcGDPpcFuture(GDPpcFuture = "SSPs-MI", unit = "constant 2017 Int$PPP")

calcGDPpcPast(GDPpcPast = "WDI-MI", unit = "constant 2017 Int$PPP")
}
\arguments{
\item{GDPFuture}{A string designating the source for the future GDP data. Available sources are:
\itemize{
\item "SSPs": IIASA \href{https://tntcat.iiasa.ac.at/SspDb/dsd?Action=htmlpage&page=welcome}{SSP database}
\item "SSP2EU": Combined SSP2 and Eurostat (for the EU countries) source
\item "SDPs":
\item "MI": Missing island dataset
}
See the "Combining data sources with '-'" section below for how to combine data sources.}

\item{unit}{A string specifying the unit of GDP. Can be either:
\itemize{
\item "constant 2017 Int$PPP" (default): Scenarios are constructed in constant 2017 Int$PPP.
\item "constant 2017 US$MER": Scenarios are constructed in constant 2017 Int$PPP and then converted with
\code{\link[GDPuc:convertGDP]{GDPuc::convertGDP()}}.
}
In all cases, GDP is returned in millions.}

\item{GDPPast}{A string designating the source for the historical GDP data. Available sources are:
\itemize{
\item "WDI": World development indicators from the World Bank
\item "MI": Missing island dataset
\item "Eurostat": Eurostat
}}

\item{GDPpcFuture}{A string designating the source for the future GDP data. Available sources are:
\itemize{
\item "SSPs":
\item "SDPs": All SDP future projections are set equal to those of SSP1.
\item "MI": Missing island dataset
}
See the "Combining data sources with '-'" section below for how to combine data sources.}

\item{GDPpcPast}{A string designating the source for the historical GDPpc data. Available sources are:
\itemize{
\item "WDI": World development indicators from the World Bank
\item "MI": Missing island dataset
}
See the "Combining data sources with '-'" section below for how to combine data sources.}
}
\description{
Get the past and future GDP scenario building blocks with calcGDPPast and calcGDPFuture, respectively.
If GDP data for a scenario is required, even if just for a single year, always use \code{\link[=calcGDP]{calcGDP()}}, as what is returned
by calcGDPPast or calcGDPFuture may not end up as is in the scenario, depending on the harmonization function.
Use calcGDPPast and calcGDPFuture only when trying to access specific GDP data.

See the "Combining data sources with '-'" section below for how to combine data sources.
}
\section{Functions}{
\itemize{
\item \code{calcGDPFuture()}: Get future GDP projections

\item \code{calcGDPpcFuture()}: Get future GDPpc projections

\item \code{calcGDPpcPast()}: Get historic GDPpc data

}}
\section{Combining data sources with "-"}{
Data sources can be combined with "-" and passed to both the pastData and futureData arguments, i.e. "WDI-MI". This
signifies that WDI data will be taken first, but missing data will be then be filled in with data from MI.
}

\keyword{internal}
