% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcGDPpc.R
\name{calcGDPpc}
\alias{calcGDPpc}
\title{calcGDPpc}
\usage{
calcGDPpc(
  GDPpcCalib = c("past_IMF_SSP", "past_IMF_SDP"),
  GDPpcPast = c("WDI", "WDI"),
  GDPpcFuture = c("SSPs", "SDPs"),
  unit = "constant 2005 Int$PPP",
  useMIData = TRUE,
  extension2150 = "bezier",
  FiveYearSteps = TRUE,
  naming = "indicator_scenario"
)
}
\arguments{
\item{GDPpcCalib}{String or vector of strings}

\item{GDPpcPast}{String or vector of strings}

\item{GDPpcFuture}{String ot vector of strings}

\item{unit}{string}

\item{useMIData}{logical}

\item{extension2150}{string, either "bezier", "constant" or "none"}

\item{FiveYearSteps}{Only five year steps if TRUE, FALSE returns years from source data}

\item{naming}{naming scheme}
}
\value{
A magpie object with sets "iso3c", "year" and "variable".
}
\description{
calcGDPpc
}
\examples{
\dontrun{
library(mrdrivers)
calcOutput("GDPpc")}

}
\seealso{
[madrat::calcOutput()

Other GDPpc functions: 
\code{\link{calcGDPpcFuture}()},
\code{\link{calcGDPpcPast}()}

Other Combined scenario functions: 
\code{\link{calcGDP}()},
\code{\link{calcPopulation}()},
\code{\link{calcUrbanPop}()},
\code{\link{calcUrban}()}
}
\concept{Combined scenario functions}
\concept{GDPpc functions}
