% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcGDPpcFuture.R
\name{calcGDPpcFuture}
\alias{calcGDPpcFuture}
\title{calcGDPpcFuture}
\usage{
calcGDPpcFuture(
  GDPpcFuture = "SSPs",
  unit = "constant 2005 Int$PPP",
  useMIData = TRUE,
  extension2150 = "bezier"
)
}
\arguments{
\item{GDPpcFuture}{String ot vector of strings}

\item{unit}{string}

\item{useMIData}{logical}

\item{extension2150}{string, either "bezier", "constant" or "none"}
}
\value{
A magpie object with sets "region", "period" and "scen".
}
\description{
calcGDPpcFuture
}
\examples{
\dontrun{
library(mrdrivers)
calcOutput("GDPpcFuture")}

}
\seealso{
\link[madrat:calcOutput]{madrat::calcOutput}

Other GDPpc functions: 
\code{\link{calcGDPpcPast}()},
\code{\link{calcGDPpc}()}
}
\concept{GDPpc functions}
