% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcGDPpcHarmonized.R
\name{calcGDPpcHarmonized}
\alias{calcGDPpcHarmonized}
\title{Get Harmonized GDPpc Data}
\usage{
calcGDPpcHarmonized(args)
}
\arguments{
\item{args}{Arguments passed on to harmonization functions}
}
\value{
magpie object with the requested output data either on country or on
regional level depending on the choice of argument "aggregate" or a list of information
if supplementary is set to TRUE.
}
\description{
Get Harmonized GDPpc Data
}
\keyword{internal}
