% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcDriver.R
\name{calcHarmonizedData}
\alias{calcHarmonizedData}
\title{Get harmonized data}
\usage{
calcHarmonizedData(driver, scenario)
}
\arguments{
\item{driver}{A string designating the driver. Available drivers are:
\itemize{
\item GDP
\item Population
\item GDPpc
\item Labour
\item Urban
}}

\item{scenario}{A string (or vector of strings) designating the scenario(s) to be returned. Use
\code{\link[=toolGetScenarioDefinition]{toolGetScenarioDefinition()}} to learn what scenarios are available.}
}
\value{
magpie object with the requested output data either on country or on
regional level depending on the choice of argument "aggregate" or a list of information
if supplementary is set to TRUE.
}
\description{
Get harmonized data
}
\seealso{
\code{\link[=calcGDPPast]{calcGDPPast()}} for the scenario builiding blocks.
}
\keyword{internal}
