% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcLabour.R
\name{calcLabour}
\alias{calcLabour}
\title{calcLabour}
\usage{
calcLabour(
  LabourFuture = c("SSPs", "SDPs", "SSP2EU"),
  extension2150 = "bezier"
)
}
\arguments{
\item{LabourFuture}{Labour future data source}

\item{extension2150}{A string specifying if/how the scenarios should be extended until 2150. Can be either:
\itemize{
\item "bezier" (default): A bezier curve extension that leads to a smooth flattening of the scenario: the
slope in the last year of the scenario is halved by 2150. Currently only works for scenarios with 2100 as
their last year.
\item "constant": The last value of the scenarios is taken as constant until 2150.
\item "none": No extension.
}}
}
\value{
A magpie object with sets "iso3c", "year" and "variable".
}
\description{
Get labour scenarios. By default the following scenarios are returned:
\itemize{
\item the SSPs, i.e. SSP1-5 and SSP2EU
\item the SDPs, i.e. SDP, SDP_EI, SDP_RC, and SDP_MC
}
}
\section{Return supplementary information}{
Set the \code{supplementary} argument of \code{\link[madrat:calcOutput]{madrat::calcOutput()}} to \code{TRUE} to return a list with the scenarios and
additional information on the unit, the data sources and the harmonization function.
}

\section{Vectorization of arguments}{
Vectors are accepted for -Calib, -Past and -Future arguments.
If given a vector, different combinations are created and returned all at once. If more than one
argument is vectorised, the arguments have to have the same length. Which time series are created can be
illustrated with the following example. Let's say the harmonization function and past data source are vectors of
length 3. Then there will be in total 3 time series that are produced: the first time series is the result of
combining the first harmonization function with the first past data source, the second time series the result of
combining the second harmonization function with the second past data source, and the third time series the result
of  using the respective third entry. The future data source used in each case is the same, since in this example
only one future data source is provided.
}

\examples{
\dontrun{
library(mrdrivers)
calcLabour()
}

}
\seealso{
\code{\link[madrat:calcOutput]{madrat::calcOutput()}}
}
\concept{Population functions}
