% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcLabourHarmonized.R
\name{calcLabourHarmonized}
\alias{calcLabourHarmonized}
\title{Get Harmonized Labour Data}
\usage{
calcLabourHarmonized(harmonization, past, future, ...)
}
\arguments{
\item{harmonization}{description}

\item{past}{description}

\item{future}{description}
}
\value{
magpie object with the requested output data either on country or on
regional level depending on the choice of argument "aggregate" or a list of information
if supplementary is set to TRUE.
}
\description{
Get Harmonized Labour Data
}
\keyword{internal}
