% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcPopulationFuture.R
\name{calcPopulationFuture}
\alias{calcPopulationFuture}
\title{calcPopulationFuture}
\usage{
calcPopulationFuture(
  PopulationFuture = "SSPs",
  useMIData = TRUE,
  extension2150 = "bezier"
)
}
\arguments{
\item{PopulationFuture}{population future data source}

\item{useMIData}{logical}

\item{extension2150}{string, either "bezier", "constant" or "none"}
}
\value{
Population in millions.
}
\description{
Calculates a time series of Pop. Different sources are available:
\itemize{ \item \code{"IIASApop"}: Source: IIASA? Lavinia?  \item
\code{"IIASApop"}: Source: Lavinia? }
}
