% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcPopulationHarmonized.R
\name{calcPopulationHarmonized}
\alias{calcPopulationHarmonized}
\title{Get Harmonized Population Data}
\usage{
calcPopulationHarmonized(harmonization, past, future, yEnd, ...)
}
\arguments{
\item{harmonization}{description}

\item{past}{description}

\item{future}{description}
}
\value{
magpie object with the requested output data either on country or on
regional level depending on the choice of argument "aggregate" or a list of information
if supplementary is set to TRUE.
}
\description{
Get Harmonized Population Data
}
\keyword{internal}
