% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcDriver.R
\name{calcScenarioConstructor}
\alias{calcScenarioConstructor}
\title{Scenario construction}
\usage{
calcScenarioConstructor(driver, scenario, popAsWeight, extension2150)
}
\arguments{
\item{driver}{A string designating the driver. Available drivers are:
\itemize{
\item GDP
\item Population
\item GDPpc
\item Labour
\item Urban
}}

\item{scenario}{A string (or vector of strings) designating the scenario(s) to be returned. Use
\code{\link[=toolGetScenarioDefinition]{toolGetScenarioDefinition()}} to learn what scenarios are available.}

\item{popAsWeight}{If TRUE, then population data of the same scenario is used as weight.}

\item{extension2150}{A string specifying if/how the scenarios should be extended until 2150. Can be either:
\itemize{
\item "bezier" (default): A bezier curve extension that leads to a smooth flattening of the scenario: the
slope in the last year of the scenario is halved by 2150. Currently only works for scenarios with 2100 as
their last year.
\item "constant": The last value of the scenarios is taken as constant until 2150.
\item "none": No extension.
}}
}
\value{
magpie object with the requested output data either on country or on
regional level depending on the choice of argument "aggregate" or a list of information
if supplementary is set to TRUE.
}
\description{
Scenario construction
}
\section{Combining data sources with "-"}{
Data sources can be combined with "-" and passed to both the pastData and futureData arguments, i.e. "WDI-MI". This
signifies that WDI data will be taken first, but missing data will be then be filled in with data from MI.
}

\seealso{
\code{\link[=calcGDPPast]{calcGDPPast()}} for the scenario builiding blocks.
}
\keyword{internal}
