% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcUrbanFuture.R
\name{calcUrbanFuture}
\alias{calcUrbanFuture}
\title{calcUrbanFuture}
\usage{
calcUrbanFuture(UrbanFuture = "SSPs", extension2150 = "none")
}
\arguments{
\item{UrbanFuture}{Urban future data source}

\item{extension2150}{A string specifying if/how the scenarios should be extended until 2150. Can be either:
\itemize{
\item "bezier" (default): A bezier curve extension that leads to a smooth flattening of the scenario: the
slope in the last year of the scenario is halved by 2150. Currently only works for scenarios with 2100 as
their last year.
\item "constant": The last value of the scenarios is taken as constant until 2150.
\item "none": No extension.
}}
}
\value{
A magpie object with sets "iso3c", "year" and "variable".
}
\description{
Calculates a time series of urban shares, using SSP projections Currently,
SSP data does not differentiate between SSPs and has some inconsistencies
with WDI in 2010
}
\examples{
\dontrun{
library(mrdrivers)
calcOutput("UrbanFuture")
}

}
\seealso{
\code{\link[madrat:calcOutput]{madrat::calcOutput()}}

Other Urban functions: 
\code{\link{calcUrbanPast}()},
\code{\link{calcUrbanPop}()},
\code{\link{calcUrban}()}
}
\concept{Urban functions}
