% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcUrbanPast.R
\name{calcUrbanPast}
\alias{calcUrbanPast}
\title{calcUrbanPast}
\usage{
calcUrbanPast(UrbanPast = "WDI")
}
\arguments{
\item{UrbanPast}{Urban past data source}
}
\value{
A magpie object with sets "iso3c", "year" and "variable".
}
\description{
Calculates a time series of urban shares
}
\examples{
\dontrun{
library(mrdrivers)
calcOutput("UrbanPast")}

}
\seealso{
\link[madrat:calcOutput]{madrat::calcOutput}

Other Urban functions: 
\code{\link{calcUrbanFuture}()},
\code{\link{calcUrbanPop}()},
\code{\link{calcUrban}()}
}
\concept{Urban functions}
