% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcUrbanPop.R
\name{calcUrbanPop}
\alias{calcUrbanPop}
\title{calcUrbanPop}
\usage{
calcUrbanPop(
  UrbanCalib = "past",
  UrbanPast = "WDI",
  UrbanFuture = c("SSPs", "SDPs", "SSP2EU"),
  extension2150 = "constant",
  FiveYearSteps = TRUE,
  naming = "indicator_scenario"
)
}
\arguments{
\item{UrbanCalib}{To what should be calibrated? past or future?}

\item{UrbanPast}{Urban past data source}

\item{UrbanFuture}{Urban future data source}

\item{extension2150}{string, either "bezier", "constant" or "none"}

\item{FiveYearSteps}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} \code{FiveYearSteps = TRUE} is no
longer supported; use the calcOutput argument \code{years}  instead, to retrieve
specific years.}

\item{naming}{naming scheme}
}
\value{
A magpie object with sets "iso3c", "year" and "variable".
}
\description{
calcUrbanPop
}
\section{Return supplementary information}{
Set the \code{supplementary} argument of \code{\link[madrat:calcOutput]{madrat::calcOutput()}} to \code{TRUE} to return a list with the scenarios and
additional information on the unit, the data sources and the harmonization function.
}

\section{Vectorization of arguments}{
The function accepts vectors for the harmonization function and past and future data sources.
If given a vector, different combinations are created and returned all at once. If more than one
argument is vectorised, the arguments have to have the same length. Which time series are created can be
illustrated with the following example. Let's say the harmonization function and past data source are vectors of
length 3. Then there will be in total 3 time series that are produced: the first time series is the result of
combining the first harmonization function with the first past data source, the second time series the result of
combining the second harmonization function with the second past data source, and the third time series the result
of  using the respective third entry. The future data source used in each case is the same, since in this example
only one future data source is provided.
}

\examples{
\dontrun{
library(mrdrivers)
calcOutput("UrbanPop")
}

}
\seealso{
\code{\link[madrat:calcOutput]{madrat::calcOutput()}}

Other Urban functions: 
\code{\link{calcUrbanFuture}()},
\code{\link{calcUrbanPast}()},
\code{\link{calcUrban}()}

Other Combined scenario functions: 
\code{\link{calcGDPpc}()},
\code{\link{calcGDP}()},
\code{\link{calcPopulation}()},
\code{\link{calcUrban}()}
}
\concept{Combined scenario functions}
\concept{Urban functions}
