% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolCompareProxyMagpie.R
\name{compare_proxy.magpie}
\alias{compare_proxy.magpie}
\title{Magpie proxy for waldo::compare}
\usage{
compare_proxy.magpie(x, path = "x")
}
\arguments{
\item{x}{x}

\item{path}{"x"}
}
\value{
list
}
\description{
Magpie proxy for waldo::compare
}
\keyword{internal}
