% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertARIADNE.R
\name{convertARIADNE}
\alias{convertARIADNE}
\title{Convert ARIADNE Reference Scenario}
\usage{
convertARIADNE(x, subtype)
}
\arguments{
\item{x}{MAgPIE object returned by readARIADNE}

\item{subtype}{data subtype. Either "population", "gdp", or "gdp_corona"}
}
\value{
magpie object of ARIADNE reference scenario data by country
}
\description{
Read ARIADNE Reference Scenario data from various .xls files as magpie object
}
\seealso{
Other ARIADNE functions: 
\code{\link{readARIADNE}()}
}
\concept{ARIADNE functions}
