% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertIMF.R
\name{convertIMF}
\alias{convertIMF}
\title{convertIMF}
\usage{
convertIMF(x, subtype = "current_account")
}
\arguments{
\item{x}{MAgPIE object returned by readIMF}

\item{subtype}{Either "current_account" or "GDPpc"}
}
\value{
magpie object of the data
}
\description{
Converts data from IMF
}
\seealso{
Other IMF functions: 
\code{\link{readIMF}()}
}
\concept{IMF functions}
