% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertJames.R
\name{convertJames}
\alias{convertJames}
\title{Convert James data}
\usage{
convertJames(x, subtype)
}
\arguments{
\item{x}{MAgPIE object returned by readJames}

\item{subtype}{String indicating the data series}
}
\value{
GDP per capita in USD05 in PPP or MER as magpie object
}
\description{
Convert James data on ISO country level.
}
\concept{James functions}
