% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertJames2019.R
\name{convertJames2019}
\alias{convertJames2019}
\title{Convert updated James data}
\usage{
convertJames2019(x, subtype)
}
\arguments{
\item{x}{MAgPIE object returned by readJames2019}

\item{subtype}{String indicating the data series}
}
\value{
GDP per capita in USD05 in PPP or MER as magpie object
}
\description{
Convert updated James data
}
\seealso{
Other James2019 functions: 
\code{\link{readJames2019}()}
}
\concept{James2019 functions}
