% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertMissingIslands.R
\name{convertMissingIslands}
\alias{convertMissingIslands}
\title{Convert MissingIslands}
\usage{
convertMissingIslands(x)
}
\arguments{
\item{x}{MAgPIE object returned by readMissingIslands}
}
\value{
Magpie object
}
\description{
Convert data from the MissingIslands dataset
}
\seealso{
Other MissingIslands functions: 
\code{\link{downloadMissingIslands}()},
\code{\link{readMissingIslands}()}
}
\concept{MissingIslands functions}
