% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertOECD.R
\name{convertOECD}
\alias{convertOECD}
\title{Convert OECD GDP / ratioPM}
\usage{
convertOECD(x, subtype)
}
\arguments{
\item{x}{MAgPIE object returned from readOECD}

\item{subtype}{data subtype. Either "gdp" or "ratioPM"}
}
\value{
magpie object of the GDP data
}
\description{
Convert GDP / ratioPM data from OECD to data on ISO country level.
}
\examples{
\dontrun{ 
readSource("OECD", "gdp")
readSource("OECD", "ratioPM")}

}
\concept{OECD functions}
