% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertPEAP.R
\name{convertPEAP}
\alias{convertPEAP}
\title{Convert Population Estimates And Projections from the World Bank}
\usage{
convertPEAP(x)
}
\arguments{
\item{x}{MAgPIE object returned by readPEAP}
}
\value{
magpie object
}
\description{
Convert data from the World Bank's Population Estimates And Projections
}
\seealso{
Other PEAP functions: 
\code{\link{readPEAP}()}
}
\concept{PEAP functions}
