% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertPWT.R
\name{convertPWT}
\alias{convertPWT}
\title{Convert PWT data}
\usage{
convertPWT(x)
}
\arguments{
\item{x}{MAgPIE object returned by readPWT}
}
\value{
Magpie object of the PWT data
}
\description{
Convert PWT data on ISO country level.
}
\seealso{
Other PWT functions: 
\code{\link{readPWT}()}
}
\concept{PWT functions}
