% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertSSP.R
\name{convertSSP}
\alias{convertSSP}
\title{Convert SSP data}
\usage{
convertSSP(x, subtype)
}
\arguments{
\item{x}{MAgPIE object returned from readSSP}

\item{subtype}{A string, either "all", "gdp", "pop", "urb", "pop2018Update",
"lab2018Update" or "ratioPM".}
}
\value{
Magpie object of the SSP data
}
\description{
Convert SSP data on ISO country level.
}
\examples{
\dontrun{
readSource("SSP", subtype = "gdp")
}

}
\concept{SSP functions}
