% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertUN_PopDiv.R
\name{convertUN_PopDiv}
\alias{convertUN_PopDiv}
\title{Convert UN Population Division Data}
\usage{
convertUN_PopDiv(x, subtype = "WPP2019_estimates")
}
\arguments{
\item{x}{\code{magclass} object containing UN Population Division data.}

\item{subtype}{String indicating version and sheet}
}
\value{
\code{magclass} object; population in millions.
}
\description{
Converts data from \code{readUN_PopDiv()} to ISO country level. "Other,
non-specified areas" is used as a stand-in for Taiwan, Province of China.
Countries missing in the data set are set to zero.
}
\seealso{
Other UN_PopDiv functions: 
\code{\link{downloadUN_PopDiv}()},
\code{\link{readUN_PopDiv}()}
}
\concept{UN_PopDiv functions}
