% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertWDI.R
\name{convertWDI}
\alias{convertWDI}
\title{Convert WDI}
\usage{
convertWDI(x, subtype)
}
\arguments{
\item{x}{MAgPIE object containing WDI data}

\item{subtype}{A string. Type of WDI data that should be read. Use the
worldbank indicator abbreviation. Available subtypes are:
\itemize{
\item \code{"SP.POP.TOTL"}: Population, total
\item \code{"NY.GDP.MKTP.PP.KD"}: GDP,PPP (constant 2017 international Dollar)
\item \code{"NY.GDP.MKTP.PP.CD"}: GDP, PPP (current international Dollar)
\item \code{"NY.GDP.MKTP.CD"}: GDP MER (current US Dollar)
\item \code{"NY.GDP.MKTP.KD"}: GDP MER (constant 2010 USDollar)
\item \code{"NY.GDP.MKTP.KN"}: GDP LCU (constant LCU)
\item \code{"NY.GDP.MKTP.CN"}: GDP LCU (current LCU)
\item \code{"NY.GDP.DEFL.KD.ZG"}: Country GDP deflator (annual \%)
\item \code{"SP.URB.TOTL.IN.ZS"}: Urban population (percentage of total)
\item \code{"NY.GDP.PCAP.CN"}:GDP, LCU, per capita (current LCU)
\item \code{"NY.GDP.PCAP.PP.KD"}: GDP PPP, per capita (2017 international $)
\item \code{"NY.GDP.PCAP.KD"}: GDP, MER, per capita (2010 US$)
\item \code{"NV.AGR.TOTL.KD"}: Ag GDP, MER, (2010 US$)
\item \code{"NV.AGR.TOTL.CD"}: Ag GDP, MER, (current US$)
\item \code{"NY.GDP.PCAP.CD"}: GDP MER, per capita(current US$)
\item \code{"NY.GDP.PCAP.PP.CD"}: GDP PPP, per capita (current int$)
\item \code{"PA.NUS.PPPC.RF"}: Price level ratio of PPP conversion factor (GDP) to market exchange rate
}}
}
\value{
A magpie object of the WDI data
}
\description{
Convert WDI converts data from readWDI() to ISO country level. Adds Taiwan
as difference from global total.
}
\seealso{
Other WDI functions: 
\code{\link{downloadWDI}()},
\code{\link{readWDI}()}
}
\concept{WDI functions}
