% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadIMF.R
\name{downloadIMF}
\alias{downloadIMF}
\title{downloadIMF}
\usage{
downloadIMF()
}
\description{
Download Wolrd Economic Outlook from the International Monetary Fund
}
\examples{
\dontrun{
downloadSource("IMF")
}
}
\seealso{
\code{\link[madrat:downloadSource]{madrat::downloadSource()}}

\code{\link[=readIMF]{readIMF()}}
}
