% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadMissingIslands.R
\name{downloadMissingIslands}
\alias{downloadMissingIslands}
\title{Download the "Missing Islands" dataset}
\usage{
downloadMissingIslands()
}
\description{
Download MissingIslands data (on population or GDP) for filling in data gaps from the
full datasets.
}
\examples{
\dontrun{
downloadSource("MissingIslands")
}
}
\seealso{
\code{\link[madrat:downloadSource]{madrat::downloadSource()}}

\code{\link[=readMissingIslands]{readMissingIslands()}}
}
