% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadUN_PopDiv.R
\name{downloadUN_PopDiv}
\alias{downloadUN_PopDiv}
\title{downloadUN_PopDiv}
\usage{
downloadUN_PopDiv()
}
\description{
Download World Population Prospects data from the United Nations Deparment on Economic and Social
Affairs
}
\examples{
\dontrun{
downloadSource("UN_PopDiv")
}
}
\seealso{
\code{\link[madrat:downloadSource]{madrat::downloadSource()}}

\code{\link[=readUN_PopDiv]{readUN_PopDiv()}}
}
