% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadWDI.R
\name{downloadWDI}
\alias{downloadWDI}
\title{Dowload WDI}
\usage{
downloadWDI()
}
\description{
Download World development indicators (WDI) data .Rds file.
the WDI data is updated with the function "WDISearch(cache = WDIcache())"
}
\examples{
\dontrun{ 
library(mrdrivers)
downloadSource("WDI")}

}
\seealso{
\url{https://databank.worldbank.org/source/world-development-indicators}, \code{\link[madrat:downloadSource]{madrat::downloadSource()}}

Other WDI functions: 
\code{\link{convertWDI}()},
\code{\link{readWDI}()}
}
\concept{WDI functions}
