% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrdrivers-package.R
\docType{package}
\name{mrdrivers-package}
\alias{mrdrivers-package}
\alias{mrdrivers}
\title{MadRat drivers Input Data Library}
\description{
Create scenarios of GDP, Population, GDP per capita, and Urban population share
}
\seealso{
Useful links:
\itemize{
  \item \url{https://pik-piam.github.io/mrdrivers}
  \item \url{https://github.com/pik-piam/mrdrivers}
  \item Report bugs at \url{https://github.com/pik-piam/mrdrivers/issues}
}

}
\author{
\strong{Maintainer}: Johannes Koch \email{jokoch@pik-potsdam.de}

Authors:
\itemize{
  \item Johannes Koch \email{jokoch@pik-potsdam.de}
  \item Bjoern Soergel
  \item Deborra Leip
  \item Falk Benke
  \item Jan Philipp Dietrich
}

}
\keyword{internal}
