% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readADB.R
\name{readADB}
\alias{readADB}
\alias{convertADB}
\alias{downloadADB}
\title{Read ADB data}
\usage{
readADB()

convertADB(x, subtype = "all")

downloadADB()
}
\arguments{
\item{x}{MAgPIE object returned from readADB}

\item{subtype}{A string, either "all", "gdppc", "pop"}
}
\value{
The read-in data, usually a magpie object. If supplementary is TRUE a list including
the data and metadata is returned instead. The temporal and data dimensionality
should match the source data. The spatial dimension should either match the source data or,
if the convert argument is set to TRUE, should be on ISO code country level.
}
\description{
Read-in an ADB data as magclass object
}
\examples{
\dontrun{
readSource("ADB", subtype = "gdppc")
}
}
\seealso{
\code{\link[madrat:readSource]{madrat::readSource()}}
}
