% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readARIADNE.R
\name{readARIADNE}
\alias{readARIADNE}
\title{Read ARIADNE Reference Scenario}
\usage{
readARIADNE(subtype)
}
\arguments{
\item{subtype}{data subtype. Either "population", "gdp", or "gdp_corona"}
}
\value{
magpie object of ARIADNE reference scenario data by country
}
\description{
Read ARIADNE Reference Scenario data from various .xls files as magpie object
}
\seealso{
\code{\link[madrat:readSource]{madrat::readSource()}}

Other ARIADNE functions: 
\code{\link{convertARIADNE}()},
\code{\link{readARIADNE_ReferenceScenario}()}
}
\concept{"Future" GDP functions}
\concept{"Future" Population functions}
\concept{ARIADNE functions}
