% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readEurostat.R
\name{readEurostat}
\alias{readEurostat}
\title{Read Eurostat historical emissions}
\usage{
readEurostat(subtype)
}
\arguments{
\item{subtype}{emissions for original eurostat emissions split, MACCemi for MACC historical emissions, or
sectorEmi for sector specific emissions}
}
\value{
magpie object of Eurostat historical emissions (MtCO2)
}
\description{
Read-in Eurostat historical emissions csv files as magclass object
}
\examples{
\dontrun{
readSource("Eurostat", subtype = "population")
}
}
\seealso{
\code{\link[madrat:readSource]{madrat::readSource()}}

Other "Past" population functions: 
\code{\link{readHYDE}()},
\code{\link{readMissingIslands}()},
\code{\link{readUN_PopDiv}()},
\code{\link{readWDI}()}

Other "Future" population functions: 
\code{\link{readPEAP}()},
\code{\link{readPopulationTWN}()}

Other "Past" GDP functions: 
\code{\link{readMissingIslands}()},
\code{\link{readWDI}()}

Other Eurostat functions: 
\code{\link{convertEurostat}()}
}
\concept{"Future" population functions}
\concept{"Past" GDP functions}
\concept{"Past" population functions}
\concept{Eurostat functions}
