% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readIIASApop.R
\name{readIIASApop}
\alias{readIIASApop}
\title{Read IIASApop}
\usage{
readIIASApop()
}
\value{
magpie object of the population data
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

Read-in an population data csv file as magclass object
}
\examples{
\dontrun{ a <- readSource(type="IIASApop")}
}
\seealso{
\code{\link{readSource}}
}
\keyword{internal}
