% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readIndiaDEA.R
\name{readIndiaDEA}
\alias{readIndiaDEA}
\alias{convertIndiaDEA}
\alias{downloadIndiaDEA}
\title{Read-in data from India's Department of Economic Affairs (DEA)}
\usage{
readIndiaDEA()

convertIndiaDEA(x, subtype = "all", subset = c("baseline", "optimistic"))

downloadIndiaDEA()
}
\arguments{
\item{x}{MAgPIE object returned from readIndiaDEA}

\item{subtype}{A string, either "all", "gdppc", "pop"}

\item{subset}{A vector of strings designating the scenarios. Defaults to c("baseline", "optimistic").}
}
\value{
The read-in data, usually a magpie object. If supplementary is TRUE a list including
the data and metadata is returned instead. The temporal and data dimensionality
should match the source data. The spatial dimension should either match the source data or,
if the convert argument is set to TRUE, should be on ISO code country level.
}
\description{
Read-in DEA data as magclass object
}
\examples{
\dontrun{
readSource("IndiaDEA", subtype = "gdppc")
}
}
\seealso{
\code{\link[madrat:readSource]{madrat::readSource()}}
}
