% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readJames.R
\name{downloadJames}
\alias{downloadJames}
\alias{readJames}
\alias{convertJames}
\title{Read-in data from the James et al. (2012) dataset}
\usage{
downloadJames()

readJames(subtype)

convertJames(x, subtype)
}
\arguments{
\item{subtype}{String indicating the data series}

\item{x}{MAgPIE object returned by readJames}
}
\value{
The read-in data, usually a magpie object. If supplementary is TRUE a list including
the data and metadata is returned instead. The temporal and data dimensionality
should match the source data. The spatial dimension should either match the source data or,
if the convert argument is set to TRUE, should be on ISO code country level.
}
\description{
Read-in GDP per-capita data from the publication James, Spencer L., Paul
Gubbins, Christopher JL Murray, and Emmanuela Gakidou. 2012. "Developing a
Comprehensive Time Series of GDP per Capita for 210 Countries from 1950 to
2015." Population Health Metrics 10 (1): 12. doi:10.1186/1478-7954-10-12.
}
\details{
The data is in Annex 3.
}
\examples{
\dontrun{
readSource("James", subtype = "gdppc")
# Or more explicitly:
readSource("James", subtype = "WB ID (2005 base year)")
}
}
\seealso{
\code{\link[madrat:readSource]{madrat::readSource()}} and \code{\link[madrat:downloadSource]{madrat::downloadSource()}}
}
