% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readMissingIslands.R
\name{readMissingIslands}
\alias{readMissingIslands}
\alias{convertMissingIslands}
\title{Read in the "Missing Islands" dataset}
\usage{
readMissingIslands(subtype)

convertMissingIslands(x)
}
\arguments{
\item{subtype}{pop for population or gdp for gdp}

\item{x}{MAgPIE object returned by readMissingIslands}
}
\value{
Magpie object
}
\description{
Read in core data for minor islands which are not included in big inventories but have a countrycode
}
\examples{
\dontrun{
readSource("MissingIslands", subtype = "pop", convert = FALSE)
}

}
\seealso{
\code{\link[madrat:readSource]{madrat::readSource()}}

\code{\link[=downloadMissingIslands]{downloadMissingIslands()}}
}
