% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readOECD.R
\name{readOECD}
\alias{readOECD}
\alias{convertOECD}
\title{Read OECD GDP or PPP2MER ratio}
\usage{
readOECD(subtype)

convertOECD(x, subtype)
}
\arguments{
\item{subtype}{data subtype. Either "gdp" or "ratioPM"}

\item{x}{MAgPIE object returned from readOECD}
}
\value{
magpie object of the GDP data
}
\description{
Read-in GDP or PPP2MER ratio data xlsx file from OECD as magclass object
}
\examples{
\dontrun{
readSource("OECD", "gdp")
readSource("OECD", "ratioPM")
}
}
\seealso{
\code{\link[madrat:readSource]{madrat::readSource()}}
}
