% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readPEAP.R
\name{downloadPEAP}
\alias{downloadPEAP}
\alias{readPEAP}
\alias{convertPEAP}
\title{Read-in data from the World Bank's Population Estimates And Projections (PEAP)}
\usage{
downloadPEAP()

readPEAP()

convertPEAP(x)
}
\arguments{
\item{x}{MAgPIE object returned by readPEAP}
}
\value{
The read-in data, usually a magpie object. If supplementary is TRUE a list including
the data and metadata is returned instead. The temporal and data dimensionality
should match the source data. The spatial dimension should either match the source data or,
if the convert argument is set to TRUE, should be on ISO code country level.
}
\description{
Read-in an csv file from the World Bank's Population Estimates And Projections (PEAP)
The PEAP data cannot seemed to be accessed by the WDI::WDI package nor the World Bank's API directly.
Manual download required from https://databank.worldbank.org/source/population-estimates-and-projections#
}
\seealso{
\code{\link[madrat:readSource]{madrat::readSource()}} and \code{\link[madrat:downloadSource]{madrat::downloadSource()}}
}
