% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readPWT.R
\name{readPWT}
\alias{readPWT}
\alias{convertPWT}
\title{Read PWT}
\usage{
readPWT()

convertPWT(x)
}
\arguments{
\item{x}{MAgPIE object returned by readPWT}
}
\value{
Magpie object of the PWT data
}
\description{
Read-in an PWT data .xlsx file as magclass object
}
\section{Functions}{
\itemize{
\item \code{convertPWT}: Convert PWT data
}}

\seealso{
Other "Past" GDPpc functions: 
\code{\link{readJames2019}()},
\code{\link{readWDI}()}
}
\concept{"Past" GDPpc functions}
