% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readUN_PopDiv.R
\name{downloadUN_PopDiv}
\alias{downloadUN_PopDiv}
\alias{readUN_PopDiv}
\alias{convertUN_PopDiv}
\title{Read UN Population Division Data}
\usage{
downloadUN_PopDiv()

readUN_PopDiv(subtype = "estimates")

convertUN_PopDiv(x)
}
\arguments{
\item{subtype}{String indicating version and sheet}

\item{x}{MAgPIE object returned from readUN_PopDiv}
}
\value{
The read-in data, usually a magpie object. If supplementary is TRUE a list including
the data and metadata is returned instead. The temporal and data dimensionality
should match the source data. The spatial dimension should either match the source data or,
if the convert argument is set to TRUE, should be on ISO code country level.
}
\description{
Read past UN population data.
}
\seealso{
\code{\link[madrat:readSource]{madrat::readSource()}} and \code{\link[madrat:downloadSource]{madrat::downloadSource()}}
}
