% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolCheckUserInput.R
\name{toolCheckUserInput}
\alias{toolCheckUserInput}
\title{Check user input}
\usage{
toolCheckUserInput(driver, args)
}
\arguments{
\item{driver}{A string designating the driver. Available drivers are:
\itemize{
\item GDP
\item Population
\item GDPpc
\item Labour
\item Urban
}}

\item{args}{A list with all the arguments passed to the parent function}
}
\value{
TRUE
}
\description{
toolCheckUserInput checks the user input.
}
\keyword{internal}
