% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolExtend2150.R
\name{toolExtend2150}
\alias{toolExtend2150}
\title{Extend until 2150 in 5 year time steps}
\usage{
toolExtend2150(data, extension2150)
}
\arguments{
\item{data}{A list with "x", a magpie object, and "description" elements}

\item{extension2150}{A string specifying if/how the scenarios should be extended until 2150. Can be either:
\itemize{
\item "bezier" (default): A bezier curve extension that leads to a smooth flattening of the scenario: the
slope in the last year of the scenario is halved by 2150. Currently only works for scenarios with 2100 as
their last year.
\item "constant": The last value of the scenarios is taken as constant until 2150.
\item "none": No extension.
}}
}
\value{
The modified data list.
}
\description{
toolExtend2150 extends a magpit object until 2150 in 5 year time steps. Either with bezierExtension or constant.
}
\keyword{internal}
