% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolGeneralConvert.R
\name{toolGeneralConvert}
\alias{toolGeneralConvert}
\title{Tool used to consolidate the most common "convert" operations}
\usage{
toolGeneralConvert(
  x,
  useDefaultSetNames = TRUE,
  countryFillWith = 0,
  substituteNAsWith = 0,
  warn = TRUE,
  note = TRUE,
  ...
)
}
\arguments{
\item{x}{A magpie object.}

\item{useDefaultSetNames}{TRUE or FALSE.}

\item{countryFillWith}{\enumerate{
\item 
}}

\item{substituteNAsWith}{\enumerate{
\item 
}}

\item{warn}{TRUE or FALSE.}

\item{note}{TRUE or FALSE.}

\item{...}{Arguments passed on to \code{\link[madrat:toolCountryFill]{madrat::toolCountryFill()}}}
}
\value{
A magpie object.
}
\description{
The most important and common "convert" operations are:
\itemize{
\item removing undefined countries,
\item using default set names "iso3c", "year", and "variable",
\item substituting NAs, see the "substituteNAsWith" argument,
\item fill in countries,
\item sort in chronological order.
}
}
