% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolGetAllDrivers.R
\name{toolGetAllDrivers}
\alias{toolGetAllDrivers}
\title{Get all drivers}
\usage{
toolGetAllDrivers(
  scenario = "SSPs",
  aggregate = TRUE,
  average2020 = FALSE,
  version = "current"
)
}
\arguments{
\item{scenario}{A string (or vector of strings) designating the scenario(s) to be returned. Use
\code{\link[=toolGetScenarioDefinition]{toolGetScenarioDefinition()}} to learn what scenarios are available.}

\item{aggregate}{Boolean indicating whether output data aggregation should be performed or
not, "GLO" (or "glo") for aggregation to one global region,
"REG+GLO" (or "regglo") for a combination of regional and global data.}

\item{average2020}{If TRUE (default), then the 2020 value is replaced by the 2018-2022 average. To be consistent,
the yearly resolution is decreased to 5 year intervals.}

\item{version}{version tag to add to the file name.}
}
\value{
A list of magpie objects
}
\description{
Return all drivers and write them to file if version not set to "current".
}
\keyword{internal}
