% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolHarmonizeFuture.R
\name{toolHarmonizeFuture}
\alias{toolHarmonizeFuture}
\title{Harmonization tool Future}
\usage{
toolHarmonizeFuture(past, future, method = "level")
}
\arguments{
\item{past}{A magpie object.}

\item{future}{A magpie object with only one scenario/datatype, i.e. the length of the third dimension should be 1.}

\item{method}{A string defining the harmonization method:
\itemize{
\item "level": the harmonized time-series is exactly equal to 'past' in the years before the first year
of 'future'.
\item "growth": the harmonized time-series follows the same growth rates as 'past', in the years
before the first year of 'future'.
}}
}
\value{
A magpie object with the same dimensions as 'past'.
}
\description{
Like all harmonization tools in mrdrivers, toolHarmonizeFuture takes two magpie objects, 'past' and 'future',
and returns a single magpie object, i.e. the harmonized time-series. In this case, the harmonized time-series is
always equal to 'future', in the years of 'future'. After that the harmonized time-series depends on the 'method'
argument chosen.
}
\section{Dimensions of 'past' and 'future'}{
If the 'past' object has multiple scenarios/datatypes, i.e. the length of the third dimension is larger than 1,
then the a harmonized time-series is created for every scenario/datatype in past. The same 'future' object is used
in every case - hence the requirement that 'future' only have one scenario/datatype.
}

\examples{
\dontrun{
toolHarmonizePast(past, future)
}
}
