% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolHarmonizePastGrFuture.R
\name{toolHarmonizePastGrFuture}
\alias{toolHarmonizePastGrFuture}
\title{Harmonization tool PastGrFuture}
\usage{
toolHarmonizePastGrFuture(past, future)
}
\arguments{
\item{past}{A magpie object with only one scenario/datatype, i.e. the length of the third dimension should be 1.}

\item{future}{A magpie object.}
}
\value{
A magpie object with the same dimensions as 'future'.
}
\description{
Like all harmonization tools in mrdrivers, toolHarmonizePastGrFuture takes two magpie objects, 'past' and 'future',
and returns a single magpie object, i.e. the harmonized time-series. In this case, the harmonized time-series is
equal to 'past', in the years of 'past', and then follows the same growth rates as 'future', for the years of
'future'.
}
\section{Dimensions of 'past' and 'future'}{
If the 'future' object has multiple scenarios/datatypes, i.e. the length of the third dimension is larger than 1,
then the a harmonized time-series is created for every scenario/datatype in future. The same 'past' object is used
in every case - hence the requirement that 'past' only have one scenario/datatype.
}

