% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolHarmonizePastTransition.R
\name{toolHarmonizePastTransition}
\alias{toolHarmonizePastTransition}
\title{Harmonization tool PastTransition}
\usage{
toolHarmonizePastTransition(past, future, yEnd, aslist = FALSE)
}
\arguments{
\item{past}{A magpie object with only one scenario/datatype, i.e. the length of the third dimension should be 1.}

\item{future}{A magpie object.}

\item{yEnd}{An integer designating the last year of the transition period.}

\item{aslist}{If TRUE (default is FALSE) then a list with the data object and a description field is returned,
insead of only the data object.}
}
\value{
A magpie object with the same dimensions as 'future'.
}
\description{
Like all harmonization tools in mrdrivers, toolHarmonizePastTransition takes two magpie objects, 'past' and 'future',
and returns a single magpie object, i.e. the harmonized time-series. In this case, the harmonized time-series is
equal to 'past', in the years of 'past', and then transitions to 'future' by the year 'yEnd'. After yEnd, the
harmonized time-series is equal to 'future'. In the transition phase, a share of the absolute difference between
'past' and future' in the last year of past' is added to future. This share starts at 1 in the last year of 'past'
and decreases linearly to 0 by yEnd.
}
\section{Dimensions of 'past' and 'future'}{
If the 'future' object has multiple scenarios/datatypes, i.e. the length of the third dimension is larger than 1,
then the a harmonized time-series is created for every scenario/datatype in future. The same 'past' object is used
in every case - hence the requirement that 'past' only have one scenario/datatype.
}

