% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolInterpolateAndExtrapolate.R
\name{toolInterpolateAndExtrapolate}
\alias{toolInterpolateAndExtrapolate}
\title{toolInterpolateAndExtrapolate}
\usage{
toolInterpolateAndExtrapolate(data, extrapolate = TRUE)
}
\arguments{
\item{data}{A magpie object}

\item{extrapolate}{TRUE or FALSE}
}
\value{
list
}
\description{
Fill in years of partially missing countries through inter- and extrapolation.
}
\keyword{internal}
