% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolFillWith.R
\name{toolListFillWith}
\alias{toolListFillWith}
\title{toolListFillWith}
\usage{
toolListFillWith(x)
}
\arguments{
\item{x}{list of objects returned by caclOutput(..., supplementary = TRUE)}
}
\value{
The completed magpie object
}
\description{
Reduce list by filling the magpie objects with the following element of the list.
}
\keyword{internal}
