% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolMbindAddMissingYears.R
\name{toolMbindAddMissingYears}
\alias{toolMbindAddMissingYears}
\title{Bind magpie objects with different temporal dimensions}
\usage{
toolMbindAddMissingYears(x, y)
}
\arguments{
\item{x}{A magpie object}

\item{y}{A magpie object}
}
\value{
A magpie object
}
\description{
Bind magpie objects even if they differ in the temporal dimension, in which any missing years are added and filled
with NAs.
}
\keyword{internal}
