% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolGetScenarioDefinition.R
\name{toolReplaceShortcuts}
\alias{toolReplaceShortcuts}
\title{Replace group shortcuts with individual scenarios}
\usage{
toolReplaceShortcuts(scenario)
}
\arguments{
\item{scenario}{A string (or vector of strings) designating the scenario(s) to be returned. Use
\code{\link[=toolGetScenarioDefinition]{toolGetScenarioDefinition()}} to learn what scenarios are available.}
}
\value{
The modified scenario argument, with the group-names replaced.
}
\description{
Some scenarios can be referred to as a group, e.g. as "SSPs". In some cases however the individual scenarios of that
group are wanted explicitly. This function replaces the group-shortcuts with the names of the scenarios in that
group.

Current available groups are: SSPs, SSP2IndiaDEAs
}
\examples{
toolReplaceShortcuts("SSPs")

}
\seealso{
\code{\link[=toolGetScenarioDefinition]{toolGetScenarioDefinition()}} for scenario options and definitions. Set \code{getGroupShortcuts} to \code{TRUE} to
return the available shortcuts: \code{toolGetScenarioDefinition(getGroupShortcuts = TRUE)}.
}
