% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolGetAllDrivers.R
\name{toolSaveAllDrivers}
\alias{toolSaveAllDrivers}
\title{Save all drivers as rds file}
\usage{
toolSaveAllDrivers(version, scenario = "SSPs")
}
\arguments{
\item{version}{version tag to add to the file name.}

\item{scenario}{A string (or vector of strings) designating the scenario(s) to be returned. Use
\code{\link[=toolGetScenarioDefinition]{toolGetScenarioDefinition()}} to learn what scenarios are available.}
}
\value{
A list of magpie objects
}
\description{
Save all drivers as rds file.
}
\keyword{internal}
