% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcDefaultDrivers.R
\name{calcDefaultDrivers}
\alias{calcDefaultDrivers}
\title{Collect Default Model Drivers}
\usage{
calcDefaultDrivers(drivers = c("Population", "Urban", "GDP", "GDPpc"))
}
\arguments{
\item{drivers}{Vector of strings.}
}
\value{
A magpie object.
}
\description{
Collect Default Model Drivers
}
