% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcCollectProjectionDrivers.R
\name{calcDefaultDrivers}
\alias{calcDefaultDrivers}
\title{Collect Default Model Drivers}
\usage{
calcDefaultDrivers(drivers = c("pop", "GDP"))
}
\arguments{
\item{drivers}{Which default drivers to collect.}
}
\description{
Collect Default Model Drivers
}
