% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcDefaultDrivers.R
\name{calcDefaultDrivers}
\alias{calcDefaultDrivers}
\title{Collect Default Model Drivers}
\usage{
calcDefaultDrivers(drivers = c("Population", "GDP", "UrbanPop"))
}
\arguments{
\item{drivers}{Vector of strings.}
}
\value{
A magpie object with sets "iso3c", "year" and "variable".
}
\description{
Collect Default Model Drivers
}
\examples{
\dontrun{
library(mrdrivers)
calcOutput("DefaultDrivers")}

}
\seealso{
\code{\link[madrat:calcOutput]{madrat::calcOutput()}}
}
