% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcDriver.R
\name{calcDriver}
\alias{calcDriver}
\title{Vectorised scenario construction}
\usage{
calcDriver(
  driver,
  scenario,
  popAsWeight = FALSE,
  naming = "indicator_scenario",
  extension2150 = "bezier",
  ...
)
}
\arguments{
\item{driver}{A string designating the driver. Available drivers are:
\itemize{
\item GDP
\item Population
\item GDPpc
\item Labour
\item Urban
}}

\item{scenario}{A string (or vector of strings) designating the scenario(s) to be returned. Use
\code{\link[=toolGetScenarioDefinition]{toolGetScenarioDefinition()}} to learn what scenarios are available.}

\item{popAsWeight}{If TRUE, then population data of the same scenario is used as weight.}

\item{naming}{A string giving the naming scheme of the data dimension. Can be either:
\itemize{
\item "indicator_scenario" (default): Returns names of the type "gdp_SSP2", or "pop_SSP2".
\item "indicator.scenario": Returns names of the type "gdp.SSP2", or "pop.SSP2".
\item "scenario": Returns names of the type "SSP2".
}
Set naming to "scenario" when you want to operate on SSP2 gdp and population data for instance, and not have to
worry about the conflicting names.}

\item{extension2150}{A string specifying if/how the scenarios should be extended until 2150. Can be either:
\itemize{
\item "bezier" (default): A bezier curve extension that leads to a smooth flattening of the scenario: the
slope in the last year of the scenario is halved by 2150. Currently only works for scenarios with 2100 as
their last year.
\item "constant": The last value of the scenarios is taken as constant until 2150.
\item "none": No extension.
}}

\item{...}{
  Arguments passed on to \code{\link[=calcScenarioConstructor]{calcScenarioConstructor}}
  \describe{
    \item{\code{harmonization}}{A string designating the harmonization function.}
    \item{\code{pastData}}{A string passed to the calc'Driver'Past function, e.g. \code{\link[=calcGDPPast]{calcGDPPast()}} or \code{\link[=calcPopulationPast]{calcPopulationPast()}}.}
    \item{\code{futureData}}{A string passed to the calc'Driver'Future function, e.g. \code{\link[=calcGDPFuture]{calcGDPFuture()}} or
\code{\link[=calcPopulationFuture]{calcPopulationFuture()}}.}
  }}
}
\value{
magpie object with the requested output data either on country or on
regional level depending on the choice of argument "aggregate" or a list of information
if supplementary is set to TRUE.
}
\description{
Internal function called by the main mrdrivers calc functions, that maps over any vectorised arguments. Should only
be of interest to package developers.
}
\section{Vectorization of arguments}{
Vectors are accepted for most arguments.
If given a vector, different combinations are created and returned all at once. If more than one
argument is vectorised, the arguments have to have the same length. Which time series are created can be
illustrated with the following example. Let's say the "harmonization" and "pastdata" arguments are vectors of
length 3. Then there will be in total 3 time series that are produced: the first time series is the result of
combining the first harmonization element with the first pastdata element, the second time series the result of
combining the second harmonization element with the second pastdata element, and the third time series the result
of using the respective third entry. The futuredata element used in each case is the same, since in this example
only one futuredata element is provided.
}

\seealso{
\itemize{
\item \code{\link[=calcGDPPast]{calcGDPPast()}}, \code{\link[=calcGDPFuture]{calcGDPFuture()}}, \code{\link[=calcGDPpcPast]{calcGDPpcPast()}} and \code{\link[=calcGDPpcFuture]{calcGDPpcFuture()}} for the builiding blocks of
the GDP and GDPpc scenarios.
\item \code{\link[=calcPopulationPast]{calcPopulationPast()}}, \code{\link[=calcPopulationFuture]{calcPopulationFuture()}}, \code{\link[=calcLabourPast]{calcLabourPast()}}, \code{\link[=calcLabourFuture]{calcLabourFuture()}}, \code{\link[=calcUrbanPast]{calcUrbanPast()}}
and \code{\link[=calcUrbanFuture]{calcUrbanFuture()}} for the builiding blocks of the Population, Labor and Urban scenarios.
}
}
\keyword{internal}
