% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcGDP.R, R/calcGDPpc.R
\name{calcGDP}
\alias{calcGDP}
\alias{calcGDPpc}
\title{Get GDP and GDP per capita scenarios}
\usage{
calcGDP(
  scenario = c("SSPs", "SDPs", "SSP2EU"),
  unit = "constant 2005 Int$PPP",
  average2020 = TRUE,
  ...
)

calcGDPpc(
  scenario = c("SSPs", "SDPs", "SSP2EU"),
  unit = "constant 2005 Int$PPP",
  average2020 = TRUE,
  ...
)
}
\arguments{
\item{scenario}{A character vector designating the scenario(s) to be returned. Use \code{\link[=toolGetScenarioDefinition]{toolGetScenarioDefinition()}} to
learn what scenarios are available.}

\item{unit}{A string specifying the unit of GDP. Can be either:
\itemize{
\item "constant 2005 Int$PPP" (default): Scenarios are constructed in constant 2005 Int$PPP.
\item "constant 2005 US$MER": Scenarios are constructed in constant 2005 Int$PPP and converted before being
returned with \code{\link[GDPuc:convertGDP]{GDPuc::convertGDP()}}.
\item "constant 2017 Int$PPP": Scenarios are constructed in constant 2017 Int$PPP.
\item "constant 2017 US$MER": Scenarios are constructed in constant 2017 Int$PPP and converted before being
returned with \code{\link[GDPuc:convertGDP]{GDPuc::convertGDP()}}.
}
In all cases, GDP is returned in millions.}

\item{average2020}{If TRUE (default), then the 2020 value is replaced by the 2018-2022 average. To be consistent,
the yearly resolution is decreased to 5 year intervals.}

\item{...}{
  Arguments passed on to \code{\link[=calcDriver]{calcDriver}}, \code{\link[=calcScenarioConstructor]{calcScenarioConstructor}}
  \describe{
    \item{\code{extension2150}}{A string specifying if/how the scenarios should be extended until 2150. Can be either:
\itemize{
\item "bezier" (default): A bezier curve extension that leads to a smooth flattening of the scenario: the
slope in the last year of the scenario is halved by 2150. Currently only works for scenarios with 2100 as
their last year.
\item "constant": The last value of the scenarios is taken as constant until 2150.
\item "none": No extension.
}}
    \item{\code{naming}}{A string giving the naming scheme of the data dimension. Can be either:
\itemize{
\item "indicator_scenario" (default): Returns names of the type "gdp_SSP2", or "pop_SSP2".
\item "indicator.scenario": Returns names of the type "gdp.SSP2", or "pop.SSP2".
\item "scenario": Returns names of the type "SSP2".
}
Set naming to "scenario" when you want to operate on SSP2 gdp and population data for instance, and not have to
worry about the conflicting names.}
    \item{\code{popAsWeight}}{If TRUE, then population data of the same scenario is used as weight.}
    \item{\code{harmonization}}{A string designating the harmonization function.}
    \item{\code{pastData}}{A string passed to the calc'Driver'Past function, e.g. \code{\link[=calcGDPPast]{calcGDPPast()}} or \code{\link[=calcPopulationPast]{calcPopulationPast()}}.}
    \item{\code{futureData}}{A string passed to the calc'Driver'Future function, e.g. \code{\link[=calcGDPFuture]{calcGDPFuture()}} or
\code{\link[=calcPopulationFuture]{calcPopulationFuture()}}.}
  }}
}
\value{
magpie object with the requested output data either on country or on
regional level depending on the choice of argument "aggregate" or a list of information
if supplementary is set to TRUE.
}
\description{
Like all scenarios in mrdrivers, the GDP and GDP per capita scenarios are the result of a harmonization exercise
between a past data and future projections. Together with the corresponding population scenarios
(see \code{\link[=calcPopulation]{calcPopulation()}}) they comprise a consisten set of scenarios. See \code{\link[=toolGetScenarioDefinition]{toolGetScenarioDefinition()}} for scenario
options and definitions. By default the following scenarios are returned:
\itemize{
\item the SSPs, i.e. SSP1-5
\item the SDPs, i.e. SDP, SDP_EI, SDP_RC, and SDP_MC
\item SSP2EU
}
}
\examples{
\dontrun{
library(mrdrivers)

# Return default scenarios
calcOutput("GDP")

# Return only the SSP2EU GDP scenario
calcOutput("GDP", scenario = "SSP2EU")

# Return the now-outdated GDP scenarios used before summer 2021,
calcOutput("GDP", scenario = "SSPsOld", extension2150 = "constant", average2020 = FALSE)
}

\dontrun{
library(mrdrivers)
calcOutput("GDPpc")
}

}
\seealso{
\itemize{
\item \code{\link[madrat:calcOutput]{madrat::calcOutput()}} for how to return supplementary information and other control options,
\item \code{\link[=calcDriver]{calcDriver()}} and \code{\link[=calcScenarioConstructor]{calcScenarioConstructor()}} for how to create new scenarios
\item \link{calcGDPPast}, \link{calcGDPFuture}, \link{calcGDPpcPast} and \link{calcGDPpcFuture} for the builiding blocks of the
scenarios.
}
}
