% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcGDP.R
\name{calcGDP}
\alias{calcGDP}
\title{calcGDP}
\usage{
calcGDP(
  GDPCalib = c("calibSSPs", "calibSDPs", "calibSSP2EU"),
  GDPPast = c("WDI-MI", "WDI-MI", "Eurostat-WDI-MI"),
  GDPFuture = c("SSPs-MI", "SDPs-MI", "SSP2EU-MI"),
  unit = "constant 2005 Int$PPP",
  extension2150 = "bezier",
  FiveYearSteps = TRUE,
  naming = "indicator_scenario"
)
}
\arguments{
\item{GDPCalib}{to what should be calibrated? past, future or a transition?}

\item{GDPPast}{GDP past data source}

\item{GDPFuture}{GDP future data source}

\item{unit}{A string. Either 'constant 2005 Int$PPP', 'constant 2005 US$MER',
'constant 2017 Int$PPP' or 'constant 2017 US$MER'. In either case, the scenario
construction is done in Int$PPP, with the conversion to US$MER coming afterward.}

\item{extension2150}{string, either "bezier", "constant" or "none"}

\item{FiveYearSteps}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} \code{FiveYearSteps = TRUE} is no
longer supported; use the calcOutput argument \code{years}  instead, to retrieve
specific years.}

\item{naming}{naming scheme}
}
\value{
A magpie object with sets "iso3c", "year" and "variable".
}
\description{
Merges time series of GDP in Purchase Power Parity (PPP) of International
Dollars of the year 2005. See \code{\link{calcGDPPast}} for past datasets, and
\code{\link{calcGDPFuture}} for future datasets. The time series are
merged via the growth rates. The first year of the future scenarios
determines the merging point. All data is calibrated specified by GDPCalib.
The extension comes after the combination of past and future, in case the
calibration method affected the future extension
}
\details{
The function accepts vectors for the harmonization function and past and future data sources.
If given a vector, different combinations are created and returned all at once. If more than one
argument is vectorised, the arguments have to have the same length. Which time series are created can be
illustrated with the following example. Let's sy the harmonization function and past data source are vetors of
length 3. Then there will be in total 3 time series that are produced: the first time series is the result of
combining the first hamonization function with the first past data source, the second time series the result of
combining the second hamonization function with the second past data source, and the third time series the result
of  using the repective third entry. The future data source used in each case is the same, since in this example
only one future data source is provided.
}
\examples{
\dontrun{
library(mrdrivers)
calcOutput("GDP")}

}
\seealso{
\code{\link[madrat:calcOutput]{madrat::calcOutput()}}

Other GDP functions: 
\code{\link{calcGDPFuture}()},
\code{\link{calcGDPPast}()}

Other Combined scenario functions: 
\code{\link{calcGDPpc}()},
\code{\link{calcPopulation}()},
\code{\link{calcUrbanPop}()},
\code{\link{calcUrban}()}
}
\concept{Combined scenario functions}
\concept{GDP functions}
