% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcGDP.R
\name{calcGDP}
\alias{calcGDP}
\alias{calcGDPpc}
\title{Get GDP and GDP per capita scenarios}
\usage{
calcGDP(
  scenario = c("SSPs", "SDPs", "SSP2EU"),
  unit = toolGetUnitDollar(inPPP = TRUE),
  average2020 = TRUE,
  ...
)

calcGDPpc(
  scenario = c("SSPs", "SDPs", "SSP2EU"),
  unit = toolGetUnitDollar(inPPP = TRUE),
  average2020 = TRUE,
  ...
)
}
\arguments{
\item{scenario}{A string (or vector of strings) designating the scenario(s) to be returned. Use
\code{\link[=toolGetScenarioDefinition]{toolGetScenarioDefinition()}} to learn what scenarios are available.}

\item{unit}{A string specifying the unit of GDP. Can be either:
\itemize{
\item "constant 2017 Int$PPP" (default):
Scenarios are constructed in constant 2017 Int$PPP.
\item "constant 2017 US$MER": Scenarios are constructed in constant 2017 Int$PPP and then converted with
\code{\link[GDPuc:convertGDP]{GDPuc::toolConvertGDP()}}.
}
In all cases, GDP is returned in millions.}

\item{average2020}{If TRUE (default), then the 2020 value is replaced by the 2018-2022 average. To be consistent,
the yearly resolution is decreased to 5 year intervals.}

\item{...}{Arguments passed on to \code{\link[=calcDriver]{calcDriver()}}, of which "extension2150" and "naming" are most often of interest.}
}
\value{
magpie object with the requested output data either on country or on
regional level depending on the choice of argument "aggregate" or a list of information
if supplementary is set to TRUE.
}
\description{
Like all scenarios in mrdrivers, the GDP and GDP per capita scenarios are the result of a harmonization exercise
between past data and future projections. Together with the corresponding population scenarios
(see \code{\link[=calcPopulation]{calcPopulation()}}) they comprise a consistent set of scenarios.

By default the following scenarios are returned:
\itemize{
\item the SSPs, i.e. SSP1-5
\item the SDPs, i.e. SDP, SDP_EI, SDP_RC, and SDP_MC
}

See the vignette: \code{vignette("scenarios")} for scenario options, definitions and references.
}
\examples{
\dontrun{
# Return default scenarios
calcOutput("GDP")
calcOutput("GDPpc")

# Return only the SSP2 GDP scenario
calcOutput("GDP", scenario = "SSP2")
}

\dontrun{
calcOutput("GDPpc")
}

}
\seealso{
\itemize{
\item \code{\link[=toolGetScenarioDefinition]{toolGetScenarioDefinition()}} for scenario options and definitions.
\item \code{\link[madrat:calcOutput]{madrat::calcOutput()}} for how to return supplementary information and other control options.
\item \code{\link[=calcDriver]{calcDriver()}}, \code{\link[=calcScenarioConstructor]{calcScenarioConstructor()}} and \code{\link[=calcHarmonizedData]{calcHarmonizedData()}} for how to create new scenarios
(for developers).
}
}
