% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcGDP.R
\name{calcGDP}
\alias{calcGDP}
\title{calcGDP}
\usage{
calcGDP(
  GDPCalib = c("past_IMF_SSP", "past_IMF_SDP", "Ariadne"),
  GDPPast = c("WDI", "WDI", "Eurostat_WDI"),
  GDPFuture = c("SSPs", "SDPs", "SSP2Ariadne"),
  unit = "constant 2005 Int$PPP",
  useMIData = TRUE,
  extension2150 = "bezier",
  FiveYearSteps = TRUE,
  naming = "indicator_scenario"
)
}
\arguments{
\item{GDPCalib}{to what should be calibrated? past, future or a transition?}

\item{GDPPast}{GDP past data source}

\item{GDPFuture}{GDP future data source}

\item{unit}{string}

\item{useMIData}{logical}

\item{extension2150}{string, either "bezier", "constant" or "none"}

\item{FiveYearSteps}{Only five year steps if TRUE, FALSE returns years from source data}

\item{naming}{naming scheme}
}
\value{
A magpie object with sets "region", "period" and "scen".
}
\description{
Merges time series of GDP in Purchase Power Parity (PPP) of International
Dollars of the year 2005. See \code{\link{calcGDPPast}} for past datasets, and
\code{\link{calcGDPFuture}} for future datasets. The time series are
merged via the growth rates. The first year of the future scenarios
determines the merging point. All data is calibrated specified by GDPCalib.
The extension comes after the combination of past and future, in case the
calibration method affected the future extension
}
\examples{
\dontrun{
library(mrdrivers)
calcOutput("GDP")}

}
\seealso{
\code{\link[madrat:calcOutput]{madrat::calcOutput()}}

Other GDP functions: 
\code{\link{calcGDPFuture}()},
\code{\link{calcGDPPast}()}

Other Combined scenario functions: 
\code{\link{calcGDPpc}()},
\code{\link{calcPopulation}()},
\code{\link{calcUrbanPop}()},
\code{\link{calcUrban}()}
}
\concept{Combined scenario functions}
\concept{GDP functions}
