% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcGDPFuture.R
\name{calcGDPFuture}
\alias{calcGDPFuture}
\title{calcGDPFuture}
\usage{
calcGDPFuture(
  GDPFuture = "SSPs",
  unit = "constant 2005 Int$PPP",
  useMIData = TRUE,
  extension2150 = "none"
)
}
\arguments{
\item{GDPFuture}{GDP future data source}

\item{unit}{string}

\item{useMIData}{logical}

\item{extension2150}{string, either "bezier", "constant" or "none"}
}
\value{
A magpie object with sets "iso3c", "year" and "variable".
}
\description{
calcGDPFuture
}
\examples{
\dontrun{
library(mrdrivers)
calcOutput("GDPFuture")}

}
\seealso{
[madrat::calcOutput()

Other GDP functions: 
\code{\link{calcGDPPast}()},
\code{\link{calcGDP}()}
}
\concept{GDP functions}
