% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcGDPFuture.R
\name{calcGDPFuture}
\alias{calcGDPFuture}
\title{calcGDPFuture}
\usage{
calcGDPFuture(
  GDPFuture = "SSPs-MI",
  unit = "constant 2005 Int$PPP",
  extension2150 = "none"
)
}
\arguments{
\item{GDPFuture}{GDP future data source}

\item{unit}{A string. Either 'constant 2005 Int$PPP', 'constant 2005 US$MER',
'constant 2017 Int$PPP' or 'constant 2017 US$MER'.}

\item{extension2150}{string, either "bezier", "constant" or "none"}
}
\value{
A magpie object with sets "iso3c", "year" and "variable".
}
\description{
calcGDPFuture
}
\examples{
\dontrun{
library(mrdrivers)
calcOutput("GDPFuture")}

}
\seealso{
\code{\link[madrat:calcOutput]{madrat::calcOutput()}}

Other GDP functions: 
\code{\link{calcGDPPast}()},
\code{\link{calcGDP}()}
}
\concept{GDP functions}
