% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcGDPFuture.R, R/calcGDPPast.R,
%   R/calcGDPpcFuture.R, R/calcGDPpcPast.R
\name{calcGDPFuture}
\alias{calcGDPFuture}
\alias{calcGDPPast}
\alias{calcGDPpcFuture}
\alias{calcGDPpcPast}
\title{Get GDP and GDPpc scenario building blocks}
\usage{
calcGDPFuture(GDPFuture = "SSPs-MI", unit = "constant 2005 Int$PPP")

calcGDPPast(GDPPast = "WDI-MI", unit = "constant 2005 Int$PPP")

calcGDPpcFuture(GDPpcFuture = "SSPsOld-MI", unit = "constant 2005 Int$PPP")

calcGDPpcPast(GDPpcPast = "WDI-MI", unit = "constant 2005 Int$PPP")
}
\arguments{
\item{GDPFuture}{A string designating the source for the future GDP data. Available sources are:
\itemize{
\item "SSPs": IIASA \href{https://tntcat.iiasa.ac.at/SspDb/dsd?Action=htmlpage&page=welcome}{SSP database}
\item "SSP2EU": Combined SSP2 and Eurostat (for the EU countries) source
\item "SDPs":
\item "MI": Missing island dataset
\item "OECD": OECD
}
See the "Combining data sources with '-'" section below for how to combine data sources.}

\item{unit}{A string specifying the unit of GDP. Can be either:
\itemize{
\item "constant 2005 Int$PPP" (default): Scenarios are constructed in constant 2005 Int$PPP.
\item "constant 2005 US$MER": Scenarios are constructed in constant 2005 Int$PPP and converted before being
returned with \code{\link[GDPuc:convertGDP]{GDPuc::convertGDP()}}.
\item "constant 2017 Int$PPP": Scenarios are constructed in constant 2017 Int$PPP.
\item "constant 2017 US$MER": Scenarios are constructed in constant 2017 Int$PPP and converted before being
returned with \code{\link[GDPuc:convertGDP]{GDPuc::convertGDP()}}.
}
In all cases, GDP is returned in millions.}

\item{GDPPast}{A string designating the source for the historical GDP data. Available sources are:
\itemize{
\item "WDI": World development indicators from the World Bank
\item "MI": Missing island dataset
\item "Eurostat": Eurostat
\item The IHME/James data set. This data set is tertiary source that uses available secondary sources (e.g. WDI
and PWT) to create completed GDP per capita time series
(See \href{https://pophealthmetrics.biomedcentral.com/articles/10.1186/1478-7954-10-12}{paper}).
To access, pass one of the subtypes
"IHME_USD05_PPP_pc", "IHME_USD05_MER_pc", "IMF_USD05_PPP_pc", "PENN_USD05_PPP_pc", "WB_USD05_PPP_pc",
"MADDISON_USD05_PPP_pc", "WB_USD05_MER_pc", "IMF_USD05_MER_pc", "UN_USD05_MER_pc". In all cases, the per
capita GDP will be multiplied by WDI population data to get absolute GDP data.
}
See the "Combining data sources with '-'" section below for how to combine data sources.}

\item{GDPpcFuture}{A string designating the source for the future GDP data. Available sources are:
\itemize{
\item "SSPs":
\item "SDPs": All SDP future projections are set equal to those of SSP1.
\item "MI": Missing island dataset
}
See the "Combining data sources with '-'" section below for how to combine data sources.}

\item{GDPpcPast}{A string designating the source for the historical GDPpc data. Available sources are:
\itemize{
\item "WDI": World development indicators from the World Bank
\item "MI": Missing island dataset
}
See the "Combining data sources with '-'" section below for how to combine data sources.}
}
\description{
Get the past and future GDP scenario building blocks with calcGDPPast and calcGDPFuture, respectively.
If GDP data for a scenario is required, even if just for a single year, always use \code{\link[=calcGDP]{calcGDP()}}, as what is returned
by calcGDPPast or calcGDPFuture may not end up as is in the scenario, depending on the harmonization function.
Use calcGDPPast and calcGDPFuture only when trying to access specific GDP data.
}
\section{Functions}{
\itemize{
\item \code{calcGDPFuture()}: Get future GDP projections

\item \code{calcGDPpcFuture()}: Get future GDPpc projections

\item \code{calcGDPpcPast()}: Get historic GDPpc data

}}
\section{Combining data sources with "-"}{
Data sources can be combined with "-" and passed to both the -Past and -Future arguments, i.e. "WDI-MI". This
signifies that WDI data will be taken first, but missing data will be then be filled in with data from MI.
}

\examples{
\dontrun{
library(mrdrivers)
calcOutput("GDPFuture")
}
\dontrun{
library(mrdrivers)
calcOutput("GDPPast")
}
\dontrun{
library(mrdrivers)
calcOutput("GDPpcFuture")
}

\dontrun{
library(mrdrivers)
calcOutput("GDPpcPast")
}

}
\keyword{internal}
