% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcGDPPast.R, R/calcGDPFuture.R,
%   R/calcPopulationFuture.R, R/calcPopulationPast.R
\name{calcGDPPast}
\alias{calcGDPPast}
\alias{calcGDPpcPast}
\alias{calcGDPFuture}
\alias{calcGDPpcFuture}
\alias{calcPopulationFuture}
\alias{calcLabourFuture}
\alias{calcUrbanFuture}
\alias{calcPopulationPast}
\alias{calcLabourPast}
\alias{calcUrbanPast}
\title{Get scenario building blocks}
\usage{
calcGDPPast(pastData = "WDI-MI-James", extension1960 = "MI-James")

calcGDPpcPast(scenario = "SSPs")

calcGDPFuture(futureData = "SSPs")

calcGDPpcFuture(scenario = "SSPs")

calcPopulationFuture(futureData = "SSPs-UN_PopDiv")

calcLabourFuture(futureData = "SSPs")

calcUrbanFuture(futureData = "SSPs")

calcPopulationPast(pastData = "WDI-UN_PopDiv-MI")

calcLabourPast(pastData = "WDI-UN_PopDiv")

calcUrbanPast(pastData = "WDI")
}
\arguments{
\item{pastData}{A string specifying the sources for historic data.}

\item{extension1960}{A string specifying how the data should be extended backwards to 1960.}

\item{scenario}{A string specifying the scenario, from which the historic GDP and population data sources are taken.}

\item{futureData}{A string specifying the sources for future projections.}
}
\description{
Get the past and future scenario building blocks.
If scenario data is required, even if just for a single year, always use the calc call to the final scenario, e.g.
\code{\link[=calcGDP]{calcGDP()}} or \code{\link[=calcPopulation]{calcPopulation()}}, as what is returned by the calc...Past and calc...Future functions may not end up
as is in the scenario, depending on the harmonization function.
These functions are only still exported for compatibility with existing code in the PIAM input data pipeline.

See the vignette: \code{vignette("scenarios")} for references for the different data sources.

See the "Combining data sources with '-'" section below for how to combine data sources.
}
\section{Combining data sources with "-"}{
Data sources can be combined with "-" and passed to both the pastData and futureData arguments, i.e. "WDI-MI". This
signifies that WDI data will be taken first, but missing data will be then be filled in with data from MI.
}

\keyword{internal}
