% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcGDPPast.R
\name{calcGDPPast}
\alias{calcGDPPast}
\title{calcGDPPast}
\usage{
calcGDPPast(GDPPast = "WDI-MI", unit = "constant 2005 Int$PPP")
}
\arguments{
\item{GDPPast}{GDP past data source}

\item{unit}{A string. Either 'constant 2005 Int$PPP', 'constant 2005 US$MER',
'constant 2017 Int$PPP' or 'constant 2017 US$MER'. In either case, the scenario
construction is done in Int$PPP, with the conversion to US$MER coming afterward.}
}
\value{
A magpie object with sets "iso3c", "year" and "variable".
}
\description{
Calculates a time series of GDP in Purchase Power Parity (PPP) of million
International Dollars of the year 2005. Different sources are available:
\itemize{ \item \code{WDI}: The PPP estimate from the World Development
Indicators (WDI) are supplemented by values for Argentina, Syria and Somalia
which are missing in the database. The values were taken from World Bank.
2014. Purchasing Power Parities and the Real Size of World Economies: A
Comprehensive Report of the 2011 International Comparison Program. The World
Bank. http://elibrary.worldbank.org/doi/book/10.1596/978-1-4648-0329-1.
table 2.13 Then, the 2011 estimate is extrapolated with the GDP projection
in local currency units (LCU), as these growth rates should be approximately
what the growth models predict. The price index from 2011 was transformed
into 2005 equivalents using the inflation rate of the United States (US), as
the PPPs are in USDollar equivalents.
\item \code{PWT}: Penn World Tables
\item \code{IHME_USD05_PPP_pc}: Publication: James, Spencer L., Paul Gubbins,
Christopher JL Murray, and Emmanuela Gakidou. 2012. "Developing a
Comprehensive Time Series of GDP per Capita for 210 Countries from 1950 to
2015."" Population Health Metrics 10 (1): 12. doi:10.1186/1478-7954-10-12.
This publication also contains further interpolated indicators that can be
used. }
}
\examples{
\dontrun{
library(mrdrivers)
calcOutput("GDPPast")
}

}
\seealso{
\code{\link[madrat:calcOutput]{madrat::calcOutput()}}

Other GDP functions: 
\code{\link{calcGDPFuture}()},
\code{\link{calcGDP}()}
}
\concept{GDP functions}
