% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcGDPpc.R
\name{calcGDPpc}
\alias{calcGDPpc}
\title{calcGDPpc}
\usage{
calcGDPpc(
  GDPpcCalib = c("calibSSPs", "calibSDPs", "calibSSP2EU"),
  GDPpcPast = c("WDI-MI", "WDI-MI", "Eurostat-WDI-MI"),
  GDPpcFuture = c("SSPs-MI", "SDPs-MI", "SSP2EU-MI"),
  unit = "constant 2005 Int$PPP",
  extension2150 = "bezier",
  FiveYearSteps = TRUE,
  average2020 = TRUE,
  naming = "indicator_scenario"
)
}
\arguments{
\item{GDPpcCalib}{String or vector of strings}

\item{GDPpcPast}{String or vector of strings}

\item{GDPpcFuture}{String or vector of strings}

\item{unit}{A string. Either 'constant 2005 Int$PPP', 'constant 2005 US$MER',
'constant 2017 Int$PPP' or 'constant 2017 US$MER'. In either case, the scenario
construction is done in Int$PPP, with the conversion to US$MER coming afterward.}

\item{extension2150}{string, either "bezier", "constant" or "none"}

\item{FiveYearSteps}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} \code{FiveYearSteps = TRUE} is no
longer supported; use the calcOutput argument \code{years}  instead, to retrieve
specific years.}

\item{average2020}{TRUE or FALSE. If TRUE, then the 2020 value is replaced by the 2018-2022 average.}

\item{naming}{naming scheme}
}
\value{
A magpie object with sets "iso3c", "year" and "variable".
}
\description{
Get GDP per capita scenarios. By default the following scenarios are returned:
\itemize{
\item the SSPs, i.e. SSP1-5 and SSP2EU
\item the SDPs, i.e. SDP, SDP_EI, SDP_RC, and SDP_MC
}

The scenarios are created by harmonizing future projections onto historical data.
}
\section{Return supplementary information}{
Set the \code{supplementary} argument of \code{\link[madrat:calcOutput]{madrat::calcOutput()}} to \code{TRUE} to return a list with the scenarios and
additional information on the unit, the data sources and the harmonization function.
}

\section{Vectorization of arguments}{
The function accepts vectors for the harmonization function and past and future data sources.
If given a vector, different combinations are created and returned all at once. If more than one
argument is vectorised, the arguments have to have the same length. Which time series are created can be
illustrated with the following example. Let's say the harmonization function and past data source are vectors of
length 3. Then there will be in total 3 time series that are produced: the first time series is the result of
combining the first harmonization function with the first past data source, the second time series the result of
combining the second harmonization function with the second past data source, and the third time series the result
of  using the respective third entry. The future data source used in each case is the same, since in this example
only one future data source is provided.
}

\examples{
\dontrun{
library(mrdrivers)
calcOutput("GDPpc")
}

}
\seealso{
\code{\link[madrat:calcOutput]{madrat::calcOutput()}}

Other GDPpc functions: 
\code{\link{calcGDPpcFuture}()},
\code{\link{calcGDPpcPast}()}

Other Combined scenario functions: 
\code{\link{calcGDP}()},
\code{\link{calcPopulation}()},
\code{\link{calcUrbanPop}()},
\code{\link{calcUrban}()}
}
\concept{Combined scenario functions}
\concept{GDPpc functions}
