% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcGDPpc.R
\name{calcGDPpc}
\alias{calcGDPpc}
\title{calcGDPpc}
\usage{
calcGDPpc(
  GDPpcCalib = c("calibSSPs", "calibSDPs", "calibSSP2EU"),
  GDPpcPast = c("WDI-MI", "WDI-MI", "Eurostat-WDI-MI"),
  GDPpcFuture = c("SSPs-MI", "SDPs-MI", "SSP2EU-MI"),
  unit = "constant 2005 Int$PPP",
  extension2150 = "bezier",
  FiveYearSteps = TRUE,
  naming = "indicator_scenario"
)
}
\arguments{
\item{GDPpcCalib}{String or vector of strings}

\item{GDPpcPast}{String or vector of strings}

\item{GDPpcFuture}{String ot vector of strings}

\item{unit}{A string. Either 'constant 2005 Int$PPP', 'constant 2005 US$MER',
'constant 2017 Int$PPP' or 'constant 2017 US$MER'.}

\item{extension2150}{string, either "bezier", "constant" or "none"}

\item{FiveYearSteps}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} \code{FiveYearSteps = TRUE} is no
longer supported; use the calcOutput argument \code{years}  instead, to retrieve
specific years.}

\item{naming}{naming scheme}
}
\value{
A magpie object with sets "iso3c", "year" and "variable".
}
\description{
calcGDPpc
}
\examples{
\dontrun{
library(mrdrivers)
calcOutput("GDPpc")
}

}
\seealso{
\code{\link[madrat:calcOutput]{madrat::calcOutput()}}

Other GDPpc functions: 
\code{\link{calcGDPpcFuture}()},
\code{\link{calcGDPpcPast}()}

Other Combined scenario functions: 
\code{\link{calcGDP}()},
\code{\link{calcPopulation}()},
\code{\link{calcUrbanPop}()},
\code{\link{calcUrban}()}
}
\concept{Combined scenario functions}
\concept{GDPpc functions}
