% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcGDPpcFuture.R
\name{calcGDPpcFuture}
\alias{calcGDPpcFuture}
\title{calcGDPpcFuture}
\usage{
calcGDPpcFuture(
  GDPpcFuture = "SSPs-MI",
  unit = "constant 2005 Int$PPP",
  extension2150 = "none"
)
}
\arguments{
\item{GDPpcFuture}{String or vector of strings}

\item{unit}{A string specifying the unit of GDP. Can be either:
\itemize{
\item "constant 2005 Int$PPP" (default): Default unit in which the scenarios are constructed.
\item "constant 2005 US$MER": Scenarios are constructed in constant 2005 Int$PPP and converted before being
returned with \link[GDPuc:convertGDP]{GDPuc::convertGDP}.
\item "constant 2017 Int$PPP": Scenarios are constructed in constant 2017 Int$PPP (See paper).
\item "constant 2017 US$MER": Scenarios are constructed in constant 2017 Int$PPP and converted before being
returned with \link[GDPuc:convertGDP]{GDPuc::convertGDP}.
}
In all cases, GDP is returned in millions.}

\item{extension2150}{A string specifying if/how the scenarios should be extended until 2150. Can be either:
\itemize{
\item "bezier" (default): A bezier curve extension that leads to a smooth flattening of the scenario: the
slope in the last year of the scenario is halved by 2150. Currently only works for scenarios with 2100 as
their last year.
\item "constant": The last value of the scenarios is taken as constant until 2150.
\item "none": No extension.
}}
}
\value{
A magpie object with sets "iso3c", "year" and "variable".
}
\description{
calcGDPpcFuture
}
\examples{
\dontrun{
library(mrdrivers)
calcOutput("GDPpcFuture")
}

}
\seealso{
\code{\link[madrat:calcOutput]{madrat::calcOutput()}}

Other GDPpc functions: 
\code{\link{calcGDPpcPast}()},
\code{\link{calcGDPpc}()}
}
\concept{GDPpc functions}
