% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcGDPpcPast.R
\name{calcGDPpcPast}
\alias{calcGDPpcPast}
\title{calcGDPpcPast}
\usage{
calcGDPpcPast(
  GDPpcPast = "WDI",
  unit = "constant 2005 Int$PPP",
  useMIData = TRUE
)
}
\arguments{
\item{GDPpcPast}{String or vector of strings}

\item{unit}{string}

\item{useMIData}{logical}
}
\value{
A magpie object with sets "region", "period" and "scen".
}
\description{
calcGDPpcPast
}
\examples{
\dontrun{
library(mrdrivers)
calcOutput("GDPpcPast")}

}
\seealso{
\link[madrat:calcOutput]{madrat::calcOutput}

Other GDPpc functions: 
\code{\link{calcGDPpcFuture}()},
\code{\link{calcGDPpc}()}
}
\concept{GDPpc functions}
