% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcGDPpcPast.R
\name{calcGDPpcPast}
\alias{calcGDPpcPast}
\title{calcGDPpcPast}
\usage{
calcGDPpcPast(GDPpcPast = "WDI-MI", unit = "constant 2005 Int$PPP")
}
\arguments{
\item{GDPpcPast}{String or vector of strings}

\item{unit}{A string specifying the unit of GDP. Can be either:
\itemize{
\item "constant 2005 Int$PPP" (default): Default unit in which the scenarios are constructed.
\item "constant 2005 US$MER": Scenarios are constructed in constant 2005 Int$PPP and converted before being
returned with \link[GDPuc:convertGDP]{GDPuc::convertGDP}.
\item "constant 2017 Int$PPP": Scenarios are constructed in constant 2017 Int$PPP (See paper).
\item "constant 2017 US$MER": Scenarios are constructed in constant 2017 Int$PPP and converted before being
returned with \link[GDPuc:convertGDP]{GDPuc::convertGDP}.
}
In all cases, GDP is returned in millions.}
}
\value{
A magpie object with sets "iso3c", "year" and "variable".
}
\description{
calcGDPpcPast
}
\examples{
\dontrun{
library(mrdrivers)
calcOutput("GDPpcPast")
}

}
\seealso{
\code{\link[madrat:calcOutput]{madrat::calcOutput()}}

Other GDPpc functions: 
\code{\link{calcGDPpcFuture}()},
\code{\link{calcGDPpc}()}
}
\concept{GDPpc functions}
