% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcDriver.R
\name{calcHarmonizedData}
\alias{calcHarmonizedData}
\title{Get Harmonized Data}
\usage{
calcHarmonizedData(driver, scenario, pastData, futureData, harmonization, ...)
}
\arguments{
\item{driver}{A string designating the driver.}

\item{scenario}{A string designating the scenario.}

\item{pastData}{A string passed to the calc'Driver'Past function, e.g. \code{\link[=calcGDPPast]{calcGDPPast()}} or \code{\link[=calcPopulationPast]{calcPopulationPast()}}.}

\item{futureData}{A string passed to the calc'Driver'Future function, e.g. \code{\link[=calcGDPFuture]{calcGDPFuture()}} or
\code{\link[=calcPopulationFuture]{calcPopulationFuture()}}.}

\item{harmonization}{A string designating the harmonization function.}

\item{...}{Arguments passed on to harmonization functions}
}
\value{
magpie object with the requested output data either on country or on
regional level depending on the choice of argument "aggregate" or a list of information
if supplementary is set to TRUE.
}
\description{
Get Harmonized Data
}
\keyword{internal}
