% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcDriver.R
\name{calcHarmonizedData}
\alias{calcHarmonizedData}
\title{Get Harmonized Data}
\usage{
calcHarmonizedData(driver, scenario, pastData, futureData, harmonization, ...)
}
\arguments{
\item{driver}{A string designating the driver. Available drivers are:
\itemize{
\item GDP
\item Population
\item GDPpc
\item Labour
\item Urban
}}

\item{scenario}{A string (or vector of strings) designating the scenario(s) to be returned. Use
\code{\link[=toolGetScenarioDefinition]{toolGetScenarioDefinition()}} to learn what scenarios are available.}

\item{pastData}{A string passed to the calc'Driver'Past function, e.g. \code{\link[=calcGDPPast]{calcGDPPast()}} or \code{\link[=calcPopulationPast]{calcPopulationPast()}}.}

\item{futureData}{A string passed to the calc'Driver'Future function, e.g. \code{\link[=calcGDPFuture]{calcGDPFuture()}} or
\code{\link[=calcPopulationFuture]{calcPopulationFuture()}}.}

\item{harmonization}{A string designating the harmonization function.}

\item{...}{Arguments passed on to harmonization functions}
}
\value{
magpie object with the requested output data either on country or on
regional level depending on the choice of argument "aggregate" or a list of information
if supplementary is set to TRUE.
}
\description{
Get Harmonized Data
}
\seealso{
\itemize{
\item \code{\link[=calcGDPPast]{calcGDPPast()}}, \code{\link[=calcGDPFuture]{calcGDPFuture()}}, \code{\link[=calcGDPpcPast]{calcGDPpcPast()}} and \code{\link[=calcGDPpcFuture]{calcGDPpcFuture()}} for the builiding blocks of
the GDP and GDPpc scenarios.
\item \code{\link[=calcPopulationPast]{calcPopulationPast()}}, \code{\link[=calcPopulationFuture]{calcPopulationFuture()}}, \code{\link[=calcLabourPast]{calcLabourPast()}}, \code{\link[=calcLabourFuture]{calcLabourFuture()}}, \code{\link[=calcUrbanPast]{calcUrbanPast()}}
and \code{\link[=calcUrbanFuture]{calcUrbanFuture()}} for the builiding blocks of the Population, Labor and Urban scenarios.
}
}
\keyword{internal}
