% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcLabour.R
\name{calcLabour}
\alias{calcLabour}
\title{calcLabour}
\usage{
calcLabour(
  LabourFuture = c("SSPs", "SDPs", "SSP2EU"),
  extension2150 = "bezier"
)
}
\arguments{
\item{LabourFuture}{Labour future data source}

\item{extension2150}{string, either "bezier", "constant" or "none"}
}
\value{
A magpie object with sets "iso3c", "year" and "variable".
}
\description{
Get Labour .
}
\examples{
\dontrun{
library(mrdrivers)
calcLabour()
}

}
\seealso{
\code{\link[madrat:calcOutput]{madrat::calcOutput()}}

Other Population functions: 
\code{\link{calcPopulationFuture}()},
\code{\link{calcPopulationPast}()},
\code{\link{calcPopulation}()}
}
\concept{Population functions}
