% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcPopulation.R
\name{calcPopulation}
\alias{calcPopulation}
\alias{calcLabour}
\alias{calcUrban}
\title{Get population and labour scenarios}
\usage{
calcPopulation(scenario, ...)

calcLabour(scenario, ...)

calcUrban(scenario, asShare = TRUE, ...)
}
\arguments{
\item{scenario}{A string (or vector of strings) designating the scenario(s) to be returned. Use
\code{\link[=toolGetScenarioDefinition]{toolGetScenarioDefinition()}} to learn what scenarios are available.}

\item{...}{
  Arguments passed on to \code{\link[=calcDriver]{calcDriver}}
  \describe{
    \item{\code{extension2150}}{A string specifying if/how the scenarios should be extended until 2150. Can be either:
\itemize{
\item "bezier" (default): A bezier curve extension that leads to a smooth flattening of the scenario: the
slope in the last year of the scenario is halved by 2150. Currently only works for scenarios with 2100 as
their last year.
\item "constant": The last value of the scenarios is taken as constant until 2150.
\item "none": No extension.
}}
  }}

\item{asShare}{If TRUE (default) urban population shares are returned. If FALSE, then urban population in millions is
returned.}
}
\value{
magpie object with the requested output data either on country or on
regional level depending on the choice of argument "aggregate" or a list of information
if supplementary is set to TRUE.
}
\description{
Like all scenarios in mrdrivers, the Population, Labour and Urban population share scenarios are the result of a
harmonization exercise between past data and future projections.

The \code{scenario} argument is used to designate the scenario(s) to be returned. Currently available Population
scenarios are: SSPs, SSP1, SSP2, SSP3, SSP4, SSP5, SDPs, ISIMIP, SSP2IndiaDEAs, SSP2IndiaMedium, SSP2IndiaHigh. See the vignette:
\code{vignette("scenarios")} and/or \code{\link[=toolGetScenarioDefinition]{toolGetScenarioDefinition()}} for more information, scenario options,
definitions and references.
}
\examples{
\dontrun{
# Return all SSP scenarios
calcOutput("Population", scenario = "SSPs")

# Return only the SSP2 GDP scenario
calcOutput("Population", scenario = "SSP2")

# Return the SSP and SDP Labour scenarios
calcOutput("Population", scenario = c("SSPs", "SDPs"))
}
\dontrun{
calcOutput("Labour", scenario = "SSPs")
}
\dontrun{
calcOutput("Urban", scenario = "SSPs")
}

}
\seealso{
\itemize{
\item \code{\link[=toolGetScenarioDefinition]{toolGetScenarioDefinition()}} for scenario options and definitions.
\item \code{\link[madrat:calcOutput]{madrat::calcOutput()}} for how to return supplementary information and other control options.
\item \code{\link[=calcDriver]{calcDriver()}}, \code{\link[=calcScenarioConstructor]{calcScenarioConstructor()}} and \code{\link[=calcHarmonizedData]{calcHarmonizedData()}} for how to create new scenarios
(for developers).
}
}
