% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcPopulation.R
\name{calcPopulation}
\alias{calcPopulation}
\title{calcPopulation}
\usage{
calcPopulation(
  PopulationCalib = c("calibSSPs", "calibSDPs", "calibSSP2EU"),
  PopulationPast = c("WDI-UN_PopDiv-MI", "WDI-UN_PopDiv-MI",
    "Eurostat-WDI-UN_PopDiv-MI"),
  PopulationFuture = c("SSPs-UN_PopDiv-MI", "SDPs-UN_PopDiv-MI", "SSP2EU-UN_PopDiv-MI"),
  extension2150 = "bezier",
  FiveYearSteps = TRUE,
  naming = "indicator_scenario"
)
}
\arguments{
\item{PopulationCalib}{A string designating the harmonization function.
Available harmonization functions are:
\itemize{
\item "calibSSPs":
\item "calibSSP2EU":
\item "calibSDPs":
\item "calibUN_PopDiv":
\item "past": deprecated
\item "future": deprecated
\item "transition": deprecated
\item "past_transition":
\item "past_grFuture":
}}

\item{PopulationPast}{A string designating the source for the historical population data.
Available sources are:
\itemize{
\item "WDI": World development indicators from the World Bank
\item "UN_PopDiv": United Nations
\item "MI": Missing island dataset
\item "Eurostat": Eurostat
}}

\item{PopulationFuture}{A string designating the source for the future population data.
Available sources are:
\itemize{
\item "SSPs":
\item "SSP2EU":
\item "SDPs":
\item "UN_PopDiv":
\item "MI":
\item "SSPs_old":
\item "SRES": deprecated
\item "IIASApop": deprecated
}}

\item{extension2150}{string, either "bezier", "constant" or "none"}

\item{FiveYearSteps}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} \code{FiveYearSteps = TRUE} is no
longer supported; use the calcOutput argument \code{years}  instead, to retrieve
specific years.}

\item{naming}{naming scheme}
}
\value{
A magpie object with sets "iso3c", "year" and "variable".
}
\description{
Create population time series by harmonizing future projections onto historical data.
}
\details{
The function accepts vectors for the harmonization function and past and future data sources.
If given a vector, different combinations are created and returned all at once. If more than one
argument is vectorised, the arguments have to have the same length. Which time series are created can be
illustrated with the following example. Let's say the harmonization function and past data source are vectors of
length 3. Then there will be in total 3 time series that are produced: the first time series is the result of
combining the first harmonization function with the first past data source, the second time series the result of
combining the second harmonization function with the second past data source, and the third time series the result
of  using the respective third entry. The future data source used in each case is the same, since in this example
only one future data source is provided.
}
\examples{
\dontrun{
library(mrdrivers)
calcOutput("GDPpc")
}

}
\seealso{
\code{\link[madrat:calcOutput]{madrat::calcOutput()}}

Other Population functions: 
\code{\link{calcLabour}()},
\code{\link{calcPopulationFuture}()},
\code{\link{calcPopulationPast}()}

Other Combined scenario functions: 
\code{\link{calcGDPpc}()},
\code{\link{calcGDP}()},
\code{\link{calcUrbanPop}()},
\code{\link{calcUrban}()}
}
\concept{Combined scenario functions}
\concept{Population functions}
