% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcPopulation.R, R/calcPopulationFuture.R,
%   R/calcPopulationPast.R
\name{calcPopulation}
\alias{calcPopulation}
\alias{calcPopulationFuture}
\alias{calcPopulationPast}
\title{Get population scenarios and building blocks}
\usage{
calcPopulation(
  scenario = c("SSPs", "SDPs", "SSP2EU"),
  PopulationCalib = NULL,
  PopulationPast = NULL,
  PopulationFuture = NULL,
  extension2150 = "bezier",
  FiveYearSteps = TRUE,
  naming = "indicator_scenario"
)

calcPopulationFuture(
  PopulationFuture = "SSPs-UN_PopDiv-MI",
  extension2150 = "none"
)

calcPopulationPast(PopulationPast = "WDI-UN_PopDiv-MI")
}
\arguments{
\item{scenario}{A character vector designating the scenario(s) to be returned. See \code{\link[=toolGetScenarioDefinition]{toolGetScenarioDefinition()}} to
learn what scenarios are available. This argument is the recommend - and quickest - way to choose
a scenario. If more fine-tunning is necessary, use the xCalib, xPast and xFuture arguments directly.}

\item{PopulationCalib}{A string designating the harmonization function.
Available harmonization functions are:
\itemize{
\item "calibSSPs": use past data from PopulationPast, then the growth rates from the Wolrld Bank's PEAP until
the final year of the IMF WEO data, and then the growth rates from PopulationFuture.
\item "calibSSP2EU":
\item "calibSDPs": same as calibSSPs
\item "calibISIMIP": use past data from PopulationPast - should be UN_PopDiv with data, currently, until 2020.
Add the 2021 projections from UN_PopDiv. Then follow the same method as past_transition.
\item "past": deprecated
\item "future": deprecated
\item "transition": deprecated
\item "past_transition":
\item "past_grFuture":
}}

\item{PopulationPast}{A string designating the source for the historical population data.
Available sources are:
\itemize{
\item "WDI": World development indicators from the World Bank
\item "UN_PopDiv": United Nations
\item "MI": Missing island dataset
\item "Eurostat": Eurostat
}
See the "Combining data sources with '-'" section below for how to combine data sources.}

\item{PopulationFuture}{A string designating the source for the future population data.
Available sources are:
\itemize{
\item "SSPs": From the Wittgenstein Center \href{http://pure.iiasa.ac.at/id/eprint/17550/}{here} and
\href{http://pure.iiasa.ac.at/id/eprint/16710/}{here}
\item "SSP2EU": Combined SSP2 and Eurostat (for the EU countries) source
\item "SDPs":
\item "UN_PopDiv": United Nations
\item "MI": Missing island dataset
\item "SSPs_old": Old SSPs from the IIASA database
\item "SRES": deprecated
\item "IIASApop": deprecated
}
See the "Combining data sources with '-'" section below for how to combine data sources.}

\item{extension2150}{A string specifying if/how the scenarios should be extended until 2150. Can be either:
\itemize{
\item "bezier" (default): A bezier curve extension that leads to a smooth flattening of the scenario: the
slope in the last year of the scenario is halved by 2150. Currently only works for scenarios with 2100 as
their last year.
\item "constant": The last value of the scenarios is taken as constant until 2150.
\item "none": No extension.
}}

\item{FiveYearSteps}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} \code{FiveYearSteps = TRUE} will be deprecated soon. Until then,
please set \code{FiveYearSteps = FALSE} and use the calcOutput argument \code{years} instead, to retrieve specific years.}

\item{naming}{A string giving the naming scheme of the data dimension. Can be either:
\itemize{
\item "indicator_scenario" (default): Returns names of the type "gdp_SSP2", or "pop_SSP2".
\item "indicator.scenario": Returns names of the type "gdp.SSP2", or "pop.SSP2".
\item "scenario": Returns names of the type "SSP2".
}
Set naming to "scenario" when you want to operate on SSP2 gdp and population data for instance, and not have to
worry about the conflicting names.}
}
\value{
A magpie object with sets "iso3c", "year" and "variable".
}
\description{
Get complete population scenarios with calcPopulation, or the past/future scenario building blocks with
calcPopulationPast and calcPopulationFuture.

Complete scenarios are created by harmonizing future projections (returned by calcPopulationFuture) onto historical
data (returned by calcPopulationPast) and cover the years between 1960 and 2100.

If population data for a scenario is required, even if just for a single year, always use calcPopulation, as what is
returned by calcPopulationPast or calcPopulationFuture may not end up as is in the scenario, depending on the
harmonization function used (see the Populationcalib argument for more information). Use calcPopulationPast and
calcPopulationFuture only when trying to access specific population data, or when constructing new
complete scenarios.

By default, calcPopulation returns the following scenarios:
\itemize{
\item the SSPs, i.e. SSP1-5
\item the SDPs, i.e. SDP, SDP_EI, SDP_RC, and SDP_MC
\item SSP2EU
}
}
\section{Functions}{
\itemize{
\item \code{calcPopulationFuture}: Get future population projections

\item \code{calcPopulationPast}: Get historic population data
}}

\section{Combining data sources with "-"}{
Past and Future data sources can be combined with "-" and given as arguments to the function, i.e. "WDI-MI". This
signifies that WDI data will be taken first, but missing data will be then be filled in with data from MI.
\subsection{Another title}{

test what happnes
}

\subsection{And another}{

anything?
}
}

\section{Vectorization of arguments}{
Vectors are accepted for -Calib, -Past and -Future arguments.
If given a vector, different combinations are created and returned all at once. If more than one
argument is vectorised, the arguments have to have the same length. Which time series are created can be
illustrated with the following example. Let's say the harmonization function and past data source are vectors of
length 3. Then there will be in total 3 time series that are produced: the first time series is the result of
combining the first harmonization function with the first past data source, the second time series the result of
combining the second harmonization function with the second past data source, and the third time series the result
of  using the respective third entry. The future data source used in each case is the same, since in this example
only one future data source is provided.
}

\section{Return supplementary information}{
Set the \code{supplementary} argument of \code{\link[madrat:calcOutput]{madrat::calcOutput()}} to \code{TRUE} to return a list with the scenarios and
additional information on the unit, the data sources and the harmonization function.
}

\examples{
\dontrun{
library(mrdrivers)
# For the default scenarios
calcOutput("Population")

# For just the SSP2EU scenario
calcOutput("Population", scenario = "SSP2EU")

# For the ISIMIP SSP scenarios
calcOutput("Population",
           scenario = "ISIMIP",
           extension2150 = "none",
           FiveYearSteps = FALSE,
           aggregate = FALSE)
}

\dontrun{
library(mrdrivers)
calcOutput("PopulationFuture")
}

\dontrun{
library(mrdrivers)
calcOutput("PopulationPast")
}

}
\seealso{
\code{\link[madrat:calcOutput]{madrat::calcOutput()}}

Other mrdrivers calc-functions: 
\code{\link{calcGDPpc}()},
\code{\link{calcGDP}()},
\code{\link{calcUrban}()}
}
\concept{mrdrivers calc-functions}
