% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcPopulation.R
\name{calcPopulation}
\alias{calcPopulation}
\alias{calcLabour}
\alias{calcUrban}
\title{Get population and labour scenarios}
\usage{
calcPopulation(scenario = c("SSPs", "SDPs", "SSP2EU"), ...)

calcLabour(scenario = c("SSPs", "SDPs", "SSP2EU"), ...)

calcUrban(scenario = c("SSPs", "SDPs", "SSP2EU"), asShare = TRUE, ...)
}
\arguments{
\item{scenario}{A string (or vector of strings) designating the scenario(s) to be returned. Use
\code{\link[=toolGetScenarioDefinition]{toolGetScenarioDefinition()}} to learn what scenarios are available.}

\item{...}{Arguments passed on to \code{\link[=calcDriver]{calcDriver()}}, of which "extension2150" and "naming" are most often of interest.
Other \code{\link[=calcDriver]{calcDriver()}} arguments are used for scenario fine-tuning and by package developers.}

\item{asShare}{If TRUE (default) urban population shares are returned. If FALSE, then urban population in millions is
returned.}
}
\value{
magpie object with the requested output data either on country or on
regional level depending on the choice of argument "aggregate" or a list of information
if supplementary is set to TRUE.
}
\description{
Like all scenarios in mrdrivers, the Population, Labour and Urban population share scenarios are the result of a
harmonization exercise between past data and future projections.

By default the following scenarios are returned:
\itemize{
\item the SSPs, i.e. SSP1-5
\item the SDPs, i.e. SDP, SDP_EI, SDP_RC, and SDP_MC
}

See the vignette: \code{vignette("scenarios")} for scenario options, definitions and references.
}
\examples{
\dontrun{
# Return the default scenarios
calcOutput("Population")

# Return only the SSP2 scenario
calcOutput("Population", scenario = "SSP2")

# Return the ISIMIP SSP scenarios
calcOutput("Population", scenario = "ISIMIP", extension2150 = "none", aggregate = FALSE)
}
\dontrun{
calcOutput("Labour")
}
\dontrun{
calcOutput("Urban")
}

}
\seealso{
\itemize{
\item \code{\link[=toolGetScenarioDefinition]{toolGetScenarioDefinition()}} for scenario options and definitions.
\item \code{\link[madrat:calcOutput]{madrat::calcOutput()}} for how to return supplementary information and other control options.
\item \code{\link[=calcDriver]{calcDriver()}}, \code{\link[=calcScenarioConstructor]{calcScenarioConstructor()}} and \code{\link[=calcHarmonizedData]{calcHarmonizedData()}} for how to create new scenarios
(for developers).
}
}
