% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcPopulation.R
\name{calcPopulation}
\alias{calcPopulation}
\title{calcPopulation}
\usage{
calcPopulation(
  PopulationCalib = c("past_grPEAP_grFuture", "past_grPEAP_grFuture", "Ariadne"),
  PopulationPast = c("WDI", "WDI", "Eurostat_WDI"),
  PopulationFuture = c("SSPs", "SDPs", "SSP2Ariadne"),
  useMIData = TRUE,
  extension2150 = "bezier",
  FiveYearSteps = TRUE,
  naming = "indicator_scenario"
)
}
\arguments{
\item{PopulationCalib}{to what should be calibrated? past, future or a transition?}

\item{PopulationPast}{population past data source}

\item{PopulationFuture}{population future data source}

\item{useMIData}{logical}

\item{extension2150}{string, either "bezier", "constant" or "none"}

\item{FiveYearSteps}{Only five year steps if TRUE, FALSE returns years from source data}

\item{naming}{naming scheme}
}
\value{
A magpie object with sets "iso3c", "year" and "variable".
}
\description{
Merges time series of population for the past and present. See
\code{\link{calcPopulationPast}} for past datasets, and
\code{\link{calcPopulationFuture}} for future datasets. The time series are
merged via the growth rates. The first year of the future scenarios
determines the merging point. All data is calibrated either to the "past" or
the "future" dataset as specified by PopulationCalib.
}
\examples{
\dontrun{
library(mrdrivers)
calcOutput("GDPpc")}

}
\seealso{
\link[madrat:calcOutput]{madrat::calcOutput}

Other Population functions: 
\code{\link{calcPopulationFuture}()},
\code{\link{calcPopulationPast}()}

Other Combined scenario functions: 
\code{\link{calcGDPpc}()},
\code{\link{calcGDP}()},
\code{\link{calcUrbanPop}()},
\code{\link{calcUrban}()}
}
\concept{Combined scenario functions}
\concept{Population functions}
