% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcPopulationFuture.R
\name{calcPopulationFuture}
\alias{calcPopulationFuture}
\title{calcPopulationFuture}
\usage{
calcPopulationFuture(
  PopulationFuture = "SSPs-UN_PopDiv-MI",
  extension2150 = "none"
)
}
\arguments{
\item{PopulationFuture}{population future data source}

\item{extension2150}{string, either "bezier", "constant" or "none"}
}
\value{
A magpie object with sets "iso3c", "year" and "variable".
}
\description{
Calculates a time series of Pop. Different sources are available:
\itemize{ \item \code{"IIASApop"}: Source: IIASA? Lavinia?  \item
\code{"IIASApop"}: Source: Lavinia? }
}
\examples{
\dontrun{
library(mrdrivers)
calcOutput("PopulationFuture")}

}
\seealso{
\code{\link[madrat:calcOutput]{madrat::calcOutput()}}

Other Population functions: 
\code{\link{calcLabour}()},
\code{\link{calcPopulationPast}()},
\code{\link{calcPopulation}()}
}
\concept{Population functions}
