% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcPopulationFuture.R
\name{calcPopulationFuture}
\alias{calcPopulationFuture}
\title{calcPopulationFuture}
\usage{
calcPopulationFuture(
  PopulationFuture = "SSPs-UN_PopDiv-MI",
  extension2150 = "none"
)
}
\arguments{
\item{PopulationFuture}{A string designating the source for the future population data.
Available sources are:
\itemize{
\item "SSPs":
\item "SSP2EU":
\item "SDPs":
\item "UN_PopDiv":
\item "MI":
\item "SSPs_old":
\item "SRES": deprecated
\item "IIASApop": deprecated
}}

\item{extension2150}{string, either "bezier", "constant" or "none"}
}
\value{
A magpie object with sets "iso3c", "year" and "variable".
}
\description{
Calculate a time series of future population data.
}
\examples{
\dontrun{
library(mrdrivers)
calcOutput("PopulationFuture")
}

}
\seealso{
\code{\link[madrat:calcOutput]{madrat::calcOutput()}}

Other Population functions: 
\code{\link{calcLabour}()},
\code{\link{calcPopulationPast}()},
\code{\link{calcPopulation}()}
}
\concept{Population functions}
