% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcLabourFuture.R, R/calcLabourPast.R,
%   R/calcPopulationFuture.R, R/calcPopulationPast.R, R/calcUrbanFuture.R,
%   R/calcUrbanPast.R
\name{calcLabourFuture}
\alias{calcLabourFuture}
\alias{calcLabourPast}
\alias{calcPopulationFuture}
\alias{calcPopulationPast}
\alias{calcUrbanFuture}
\alias{calcUrbanPast}
\title{Get Population and Labour scenario building blocks}
\usage{
calcLabourFuture(LabourFuture = "SSPs")

calcLabourPast(LabourPast = "WDI")

calcPopulationFuture(PopulationFuture = "SSPs-UN_PopDiv-MI")

calcPopulationPast(PopulationPast = "WDI-UN_PopDiv-MI")

calcUrbanFuture(UrbanFuture = "SSPs")

calcUrbanPast(UrbanPast = "WDI")
}
\arguments{
\item{LabourFuture}{A string designating the source for the future working-age population data.
Available sources are:
\itemize{
\item "SSPs":
}}

\item{LabourPast}{A string designating the source for the historical working-age population data.
Available sources are:
\itemize{
\item "WDI": World development indicators from the World Bank
}}

\item{PopulationFuture}{A string designating the source for the future population data.
Available sources are:
\itemize{
\item "SSPs": From the Wittgenstein Center \href{http://pure.iiasa.ac.at/id/eprint/17550/}{here} and
\href{http://pure.iiasa.ac.at/id/eprint/16710/}{here}
\item "SSP2EU": Combined SSP2 and Eurostat (for the EU countries) source
\item "SDPs":
\item "UN_PopDiv": United Nations
\item "MI": Missing island dataset
}}

\item{PopulationPast}{A string designating the source for the historical population data.
Available sources are:
\itemize{
\item "WDI": World development indicators from the World Bank
\item "UN_PopDiv": United Nations
\item "MI": Missing island dataset
\item "Eurostat": Eurostat
}}

\item{UrbanFuture}{A string designating the source for the future urban population-share data.
Available sources are:
\itemize{
\item "SSPs":
\item "SDPs":
\item "SSP2EU":
}}

\item{UrbanPast}{A string designating the source for the historical urban population-share data.
Available sources are:
\itemize{
\item "WDI": World development indicators from the World Bank
}}
}
\description{
See the "Combining data sources with '-'" section below for how to combine data sources.
}
\section{Combining data sources with "-"}{
Data sources can be combined with "-" and passed to both the pastData and futureData arguments, i.e. "WDI-MI". This
signifies that WDI data will be taken first, but missing data will be then be filled in with data from MI.
}

\keyword{internal}
