% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcPopulationPast.R
\name{calcPopulationPast}
\alias{calcPopulationPast}
\title{calcPopulationPast}
\usage{
calcPopulationPast(PopulationPast = "WDI", useMIData = TRUE)
}
\arguments{
\item{PopulationPast}{population past data source}

\item{useMIData}{logical}
}
\value{
A magpie object with sets "iso3c", "year" and "variable".
}
\description{
Calculates a time series of Population. Different sources are available:
\itemize{
\item \code{WDI}: Source: Worldbank. Taiwan was estimated as the
difference between all countries and the global total.
\item \code{UN_PopDiv}: UN Population Division data. Taiwan is estimated
from "other, non-specified areas". Missing countries have their
values set to zero.
}
}
\examples{
\dontrun{
library(mrdrivers)
calcOutput("PopulationPast")}

}
\seealso{
\link[madrat:calcOutput]{madrat::calcOutput}

Other Population functions: 
\code{\link{calcPopulationFuture}()},
\code{\link{calcPopulation}()}
}
\concept{Population functions}
