% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcPopulationPast.R
\name{calcPopulationPast}
\alias{calcPopulationPast}
\title{calcPopulationPast}
\usage{
calcPopulationPast(PopulationPast = "WDI-UN_PopDiv-MI")
}
\arguments{
\item{PopulationPast}{A string designating the source for the historical population data.
Available sources are:
\itemize{
\item "WDI": World development indicators from the World Bank
\item "UN_PopDiv": United Nations
\item "MI": Missing island dataset
\item "Eurostat": Eurostat
}}
}
\value{
A magpie object with sets "iso3c", "year" and "variable".
}
\description{
Calculate a time series of past population data.
}
\examples{
\dontrun{
library(mrdrivers)
calcOutput("PopulationPast")
}

}
\seealso{
\code{\link[madrat:calcOutput]{madrat::calcOutput()}}

Other Population functions: 
\code{\link{calcLabour}()},
\code{\link{calcPopulationFuture}()},
\code{\link{calcPopulation}()}
}
\concept{Population functions}
