% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcRatioPPP2MER.R
\name{calcRatioPPP2MER}
\alias{calcRatioPPP2MER}
\title{Get PPP to MER ratio}
\usage{
calcRatioPPP2MER(from = "WDI", when = 2005)
}
\arguments{
\item{from}{A string indicating the source}
}
\description{
Get PPP to MER ratio
}
\examples{
\dontrun{
library(mrdrivers)
calcOutput("calcfrom")}

}
\seealso{
\code{\link[madrat:calcOutput]{madrat::calcOutput()}}
}
