% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcRatioPPP2MER.R
\name{calcRatioPPP2MER}
\alias{calcRatioPPP2MER}
\title{MER over PPP ratio}
\usage{
calcRatioPPP2MER(when = 2017)
}
\arguments{
\item{when}{An integer (defaults to 2017) specifying the year of the PPP2MER factor.}
}
\value{
magpie object with the requested output data either on country or on
regional level depending on the choice of argument "aggregate" or a list of information
if supplementary is set to TRUE.
}
\description{
Get a conversion factor to convert GDP in constant 2017 Int$PPP into constant 2017 US$MER.
Use the "when" argument to switch the year of the conversion factor. Source = WDI. Countries with missing data are
filled in with 1. Regional aggregation is weighed by GDP from WDI-MI in the year set by "when".
}
\examples{
\dontrun{
calcOutput("RatioPPP2MER")
}

}
\seealso{
\code{\link[madrat:calcOutput]{madrat::calcOutput()}}
}
