% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcUrbanFuture.R
\name{calcUrbanFuture}
\alias{calcUrbanFuture}
\title{calcUrbanFuture}
\usage{
calcUrbanFuture(UrbanFuture = "SSPs", extension2150 = "constant")
}
\arguments{
\item{UrbanFuture}{Urban future data source}

\item{extension2150}{string, either "bezier", "constant" or "none"}
}
\value{
Urban shares
}
\description{
Calculates a time series of urban shares, using SSP projections Currently,
SSP data does not differentiate between SSPs and has some unconsistencies
with WDI in 2010
}
