% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcUrbanFuture.R
\name{calcUrbanFuture}
\alias{calcUrbanFuture}
\title{calcUrbanFuture}
\usage{
calcUrbanFuture(UrbanFuture = "SSPs", extension2150 = "none")
}
\arguments{
\item{UrbanFuture}{Urban future data source}

\item{extension2150}{string, either "bezier", "constant" or "none"}
}
\value{
A magpie object with sets "iso3c", "year" and "variable".
}
\description{
Calculates a time series of urban shares, using SSP projections Currently,
SSP data does not differentiate between SSPs and has some unconsistencies
with WDI in 2010
}
\examples{
\dontrun{
library(mrdrivers)
calcOutput("UrbanFuture")}

}
\seealso{
\link[madrat:calcOutput]{madrat::calcOutput}

Other Urban functions: 
\code{\link{calcUrbanPast}()},
\code{\link{calcUrbanPop}()},
\code{\link{calcUrban}()}
}
\concept{Urban functions}
