% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcUrbanPop.R
\name{calcUrbanPop}
\alias{calcUrbanPop}
\title{calcUrbanPop}
\usage{
calcUrbanPop(
  UrbanCalib = "past",
  UrbanPast = "WDI",
  UrbanFuture = "SSPs",
  extension2150 = "constant",
  FiveYearSteps = TRUE,
  naming = "indicator_scenario"
)
}
\arguments{
\item{UrbanCalib}{To what should be calibrated? past or future?}

\item{UrbanPast}{Urban past data source}

\item{UrbanFuture}{Urban future data source}

\item{extension2150}{string, either "bezier", "constant" or "none"}

\item{FiveYearSteps}{Only five year steps if TRUE, FALSE returns years from source data}

\item{naming}{naming scheme}
}
\value{
A magpie object with sets "region", "period" and "scen".
}
\description{
calcUrbanPop
}
\examples{
\dontrun{
library(mrdrivers)
calcOutput("UrbanPop")}

}
\seealso{
\link[madrat:calcOutput]{madrat::calcOutput}

Other Urban functions: 
\code{\link{calcUrbanFuture}()},
\code{\link{calcUrbanPast}()},
\code{\link{calcUrban}()}

Other Combined scenario functions: 
\code{\link{calcGDPpc}()},
\code{\link{calcGDP}()},
\code{\link{calcPopulation}()},
\code{\link{calcUrban}()}
}
\concept{Combined scenario functions}
\concept{Urban functions}
