% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcUrbanPop.R
\name{calcUrbanPop}
\alias{calcUrbanPop}
\title{calcUrbanPop}
\usage{
calcUrbanPop(
  UrbanCalib = "past",
  UrbanPast = "WDI",
  UrbanFuture = c("SSPs", "SDPs", "SSP2EU"),
  extension2150 = "constant",
  FiveYearSteps = TRUE,
  naming = "indicator_scenario"
)
}
\arguments{
\item{UrbanCalib}{To what should be calibrated? past or future?}

\item{UrbanPast}{Urban past data source}

\item{UrbanFuture}{Urban future data source}

\item{extension2150}{string, either "bezier", "constant" or "none"}

\item{FiveYearSteps}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} \code{FiveYearSteps = TRUE} is no
longer supported; use the calcOutput argument \code{years}  instead, to retrieve
specific years.}

\item{naming}{naming scheme}
}
\value{
A magpie object with sets "iso3c", "year" and "variable".
}
\description{
calcUrbanPop
}
\examples{
\dontrun{
library(mrdrivers)
calcOutput("UrbanPop")}

}
\seealso{
\code{\link[madrat:calcOutput]{madrat::calcOutput()}}

Other Urban functions: 
\code{\link{calcUrbanFuture}()},
\code{\link{calcUrbanPast}()},
\code{\link{calcUrban}()}

Other Combined scenario functions: 
\code{\link{calcGDPpc}()},
\code{\link{calcGDP}()},
\code{\link{calcPopulation}()},
\code{\link{calcUrban}()}
}
\concept{Combined scenario functions}
\concept{Urban functions}
