% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadMissingIslands.R
\name{downloadMissingIslands}
\alias{downloadMissingIslands}
\title{downloadMissingIslands}
\usage{
downloadMissingIslands()
}
\description{
Download MissingIslands data (on population or GDP) for filling in data gaps from the
full datasets.
}
\examples{
\dontrun{
downloadSource("MissingIslands")
}
}
\seealso{
\code{\link[=downloadSource]{downloadSource()}}

Other MissingIslands functions: 
\code{\link{convertMissingIslands}()},
\code{\link{readMissingIslands}()}
}
\concept{MissingIslands functions}
