% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadWDI.R
\name{downloadWDI}
\alias{downloadWDI}
\title{Download WDI}
\usage{
downloadWDI()
}
\description{
Download World development indicators (WDI) data to .Rds file.
Data is downloaded from 1960 until "current year - 1".
The WDI data is updated with the function "WDISearch(cache = WDIcache())"
}
\examples{
\dontrun{
library(mrdrivers)
downloadSource("WDI")
}

}
\seealso{
\url{https://databank.worldbank.org/source/world-development-indicators}

\code{\link[madrat:downloadSource]{madrat::downloadSource()}}

\code{\link[=readWDI]{readWDI()}}
}
