% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrdrivers-package.R
\docType{package}
\name{mrdrivers-package}
\alias{mrdrivers-package}
\alias{mrdrivers}
\title{MadRat drivers Input Data Library}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Create scenarios of GDP, Population, GDP per capita, and Urban population share
}
\keyword{internal}
