% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readARIADNE.R, R/convertARIADNE.R
\name{readARIADNE}
\alias{readARIADNE}
\alias{convertARIADNE}
\alias{readARIADNE_ReferenceScenario}
\title{Read ARIADNE Reference Scenario}
\usage{
readARIADNE(subtype)

convertARIADNE(x, subtype)

readARIADNE_ReferenceScenario(subtype)
}
\arguments{
\item{subtype}{data subtype. Either "population", "gdp", or "gdp_corona"}

\item{x}{MAgPIE object returned by readARIADNE}
}
\value{
magpie object of ARIADNE reference scenario data by country
}
\description{
Read ARIADNE Reference Scenario data from various .xls files as magpie object
}
\section{Functions}{
\itemize{
\item \code{readARIADNE_ReferenceScenario()}: Alias to \code{\link[=readARIADNE]{readARIADNE()}}. May be deprecated in the future.

}}
\seealso{
\code{\link[madrat:readSource]{madrat::readSource()}}
}
