% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readARIADNE.R
\name{readARIADNE}
\alias{readARIADNE}
\alias{convertARIADNE}
\title{Read ARIADNE Reference Scenario}
\usage{
readARIADNE(subtype)

convertARIADNE(x, subtype)
}
\arguments{
\item{subtype}{data subtype. Either "population", "gdp", or "gdp_corona"}

\item{x}{MAgPIE object returned by readARIADNE}
}
\value{
magpie object of ARIADNE reference scenario data by country
}
\description{
Read ARIADNE Reference Scenario data from various .xls files as magpie object
}
\seealso{
\code{\link[madrat:readSource]{madrat::readSource()}}
}
