% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readEurostatPopGDP.R, R/convertEurostatPopGDP.R
\name{readEurostatPopGDP}
\alias{readEurostatPopGDP}
\alias{convertEurostatPopGDP}
\title{Read Eurostat historical emissions}
\usage{
readEurostatPopGDP(subtype)

convertEurostatPopGDP(x, subtype)
}
\arguments{
\item{subtype}{emissions for original eurostat emissions split, MACCemi for MACC historical emissions, or
sectorEmi for sector specific emissions}

\item{x}{MAgPIE object returned by readEurostatPopGDP}
}
\value{
magpie object of Eurostat historical emissions (MtCO2)
}
\description{
Read-in Eurostat historical emissions csv files as magclass object
}
\examples{
\dontrun{
readSource("EurostatPopGDP", subtype = "population")
}
}
\seealso{
\code{\link[madrat:readSource]{madrat::readSource()}}
}
