% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readIMF.R
\name{downloadIMF}
\alias{downloadIMF}
\alias{readIMF}
\alias{convertIMF}
\title{Read-in data from the International Monetary Fund's (IMF) Economic Outlook}
\usage{
downloadIMF()

readIMF()

convertIMF(x, subtype = "all")
}
\arguments{
\item{x}{MAgPIE object returned by readIMF}

\item{subtype}{Use to filter the IMF data}
}
\value{
The read-in data, usually a magpie object. If supplementary is TRUE a list including
the data and metadata is returned instead. The temporal and data dimensionality
should match the source data. The spatial dimension should either match the source data or,
if the convert argument is set to TRUE, should be on ISO code country level.
}
\description{
Read-in data from the IMF's World Economic Outlook.
Currently reading GDP per capita (in constant 2021 Int$PPP) and current account balance data.
}
\examples{
\dontrun{
readSource("IMF")
}
}
\seealso{
\code{\link[madrat:readSource]{madrat::readSource()}} and \code{\link[madrat:downloadSource]{madrat::downloadSource()}}
}
