% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readIMF.R
\name{downloadIMF}
\alias{downloadIMF}
\alias{readIMF}
\alias{convertIMF}
\title{Read IMF}
\usage{
downloadIMF()

readIMF(subtype = "current_account", subset = "WEOOct2022all.xls")

convertIMF(x, subtype = "current_account")
}
\arguments{
\item{subtype}{Either "current_account" or "GDPpc"}

\item{subset}{Either "WEOOct2022all.xls" or "WEOallOct2019.xls"}

\item{x}{MAgPIE object returned by readIMF}
}
\value{
The read-in data, usually a magpie object. If supplementary is TRUE a list including
the data and metadata is returned instead. The temporal and data dimensionality
should match the source data. The spatial dimension should either match the source data or,
if the convert argument is set to TRUE, should be on ISO code country level.
}
\description{
Read-in IMF data
}
\examples{
\dontrun{
readSource("IMF")
}
}
\seealso{
\code{\link[madrat:readSource]{madrat::readSource()}} and \code{\link[madrat:downloadSource]{madrat::downloadSource()}}
}
