% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readIMF.R
\name{readIMF}
\alias{readIMF}
\alias{convertIMF}
\title{Read IMF}
\usage{
readIMF(subtype = "current_account", subset = "WEOall.xls")

convertIMF(x, subtype = "current_account")
}
\arguments{
\item{subtype}{Either "current_account" or "GDPpc"}

\item{subset}{Either "WEOall.xls" or "WEOallOct2019.xls"}

\item{x}{MAgPIE object returned by readIMF}
}
\value{
magpie object of the data
}
\description{
Read-in IMF data
}
\section{Functions}{
\itemize{
\item \code{convertIMF()}: Convert IMF data

}}
\examples{
\dontrun{
a <- readSource(type = "IMF")
}
}
\seealso{
\code{\link[madrat:readSource]{madrat::readSource()}}

\code{\link[=downloadIMF]{downloadIMF()}}
}
