% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readIMF.R
\name{readIMF}
\alias{readIMF}
\title{Read IMF}
\usage{
readIMF(subtype = "current_account")
}
\arguments{
\item{subtype}{Either "current_account" or "GDPpc"}
}
\value{
magpie object of the data
}
\description{
Read-in iMF data
}
\examples{
\dontrun{
a <- readSource(type = "IMF")
}
}
\seealso{
\code{\link[madrat:readSource]{madrat::readSource()}}

Other "Past" GDPpc functions: 
\code{\link{readJames2019}()},
\code{\link{readPWT}()},
\code{\link{readWDI}()}

Other IMF functions: 
\code{\link{convertIMF}()}
}
\concept{"Future" GDPpc functions}
\concept{"Past" GDPpc functions}
\concept{IMF functions}
