% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readMissingIslands.R
\name{readMissingIslands}
\alias{readMissingIslands}
\title{readMissingIslands}
\usage{
readMissingIslands(subtype)
}
\arguments{
\item{subtype}{pop for population or gdp for gdp}
}
\value{
Magpie object
}
\description{
Read in core data for minor islands which are not included in big inventories but have a countrycode
}
\examples{
\dontrun{ 
readSource("MissingIslands", subtype = "pop", convert = FALSE)
}

}
\author{
Benjamin Bodirsky, Eweron Araujo
}
