% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readMissingIslands.R
\name{readMissingIslands}
\alias{readMissingIslands}
\alias{convertMissingIslands}
\alias{downloadMissingIslands}
\title{Read in the "Missing Islands" dataset}
\usage{
readMissingIslands(subtype)

convertMissingIslands(x)

downloadMissingIslands()
}
\arguments{
\item{subtype}{pop for population, or gdp for gdp}

\item{x}{MAgPIE object returned by readMissingIslands}
}
\value{
Magpie object
}
\description{
Read in gdp or population data for minor islands (not included in big inventories) from a custom made
data set that gets data from a variety of sources (e.g. CIA World Factbook, Insee, BEA, PRISM, and Woldometers).
}
\examples{
\dontrun{
readSource("MissingIslands", subtype = "pop", convert = FALSE)
}

}
\seealso{
\code{\link[madrat:readSource]{madrat::readSource()}}
}
