% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readMissingIslands.R
\name{downloadMissingIslands}
\alias{downloadMissingIslands}
\alias{readMissingIslands}
\alias{convertMissingIslands}
\title{Read in the "Missing Islands" dataset}
\usage{
downloadMissingIslands()

readMissingIslands(subtype)

convertMissingIslands(x)
}
\arguments{
\item{subtype}{pop for population, or gdp for gdp}

\item{x}{MAgPIE object returned by readMissingIslands}
}
\value{
The read-in data, usually a magpie object. If supplementary is TRUE a list including
the data and metadata is returned instead. The temporal and data dimensionality
should match the source data. The spatial dimension should either match the source data or,
if the convert argument is set to TRUE, should be on ISO code country level.
}
\description{
Read in gdp or population data for minor islands (not included in big inventories) from a custom made
data set that gets data from a variety of sources (e.g. CIA World Factbook, Insee, BEA, PRISM, and Woldometers).
}
\examples{
\dontrun{
readSource("MissingIslands", subtype = "pop", convert = FALSE)
}
}
\seealso{
\code{\link[madrat:readSource]{madrat::readSource()}} and \code{\link[madrat:downloadSource]{madrat::downloadSource()}}
}
