% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readPEAP.R
\name{readPEAP}
\alias{readPEAP}
\title{Read Population Estimates And Projections from the World Bank}
\usage{
readPEAP()
}
\value{
magpie object
}
\description{
Read-in xlsx file from the World Bank's Population Estimates And Projections
}
\seealso{
\code{\link[madrat:readSource]{madrat::readSource()}}

Other "Future" population functions: 
\code{\link{readEurostat}()},
\code{\link{readPopulationTWN}()}
}
\concept{"Future" population functions}
\concept{PEAP functions}
