% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readPWT.R
\name{readPWT}
\alias{readPWT}
\title{Read PWT}
\usage{
readPWT()
}
\value{
Magpie object of the PWT data
}
\description{
Read-in an PWT data .xlsx file as magclass object
}
\seealso{
Other "Past" GDPpc functions: 
\code{\link{readIMF}()},
\code{\link{readJames2019}()},
\code{\link{readWDI}()}

Other PWT functions: 
\code{\link{convertPWT}()}
}
\concept{"Past" GDPpc functions}
\concept{PWT functions}
