% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readSSP.R
\name{readSSP}
\alias{readSSP}
\alias{convertSSP}
\title{Read SSP}
\usage{
readSSP(subtype)

convertSSP(x, subtype)
}
\arguments{
\item{subtype}{A string, either "all", "gdp", "pop", "urb", "pop2018Update",
"lab2018Update" or "ratioPM".}

\item{x}{MAgPIE object returned from readSSP}
}
\value{
The read-in data, usually a magpie object. If supplementary is TRUE a list including
the data and metadata is returned instead. The temporal and data dimensionality
should match the source data. The spatial dimension should either match the source data or,
if the convert argument is set to TRUE, should be on ISO code country level.
}
\description{
Read-in an SSP data csv.zip file as magclass object
}
\examples{
\dontrun{
readSource("SSP", subtype = "gdp")
}
}
\seealso{
\code{\link[madrat:readSource]{madrat::readSource()}}
}
