% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readUN_PopDiv.R
\name{readUN_PopDiv}
\alias{readUN_PopDiv}
\title{Read UN Population Division Data}
\usage{
readUN_PopDiv(subtype = "WPP2019_estimates")
}
\arguments{
\item{subtype}{String indicating version and sheet}
}
\value{
\code{magclass} object; population in thousands.
}
\description{
Read past UN population data. Covers 1950 to 2015 yearly and per M.49 area.
See \emph{United Nations, Department of Economic and Social Affairs,
Population Division} "World Population Prospects: The 2015 Revision"
(\href{https://esa.un.org/unpd/wpp/}{website}).
}
\seealso{
\code{\link[madrat:readSource]{madrat::readSource()}}

Other "Past" population functions: 
\code{\link{readEurostatPopGDP}()},
\code{\link{readMissingIslands}()},
\code{\link{readWDI}()}

Other UN_PopDiv functions: 
\code{\link{convertUN_PopDiv}()},
\code{\link{downloadUN_PopDiv}()}
}
\concept{"Past" population functions}
\concept{UN_PopDiv functions}
