% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertUN_PopDiv.R, R/readUN_PopDiv.R
\name{convertUN_PopDiv}
\alias{convertUN_PopDiv}
\alias{readUN_PopDiv}
\title{Read UN Population Division Data}
\usage{
convertUN_PopDiv(x, subtype = "WPP2019_estimates")

readUN_PopDiv(subtype = "WPP2019_estimates")
}
\arguments{
\item{x}{MAgPIE object returned from readUN_PopDiv}

\item{subtype}{String indicating version and sheet}
}
\value{
\code{magclass} object; population in thousands.
}
\description{
Read past UN population data. Covers 1950 to 2015 yearly and per M.49 area.
See \emph{United Nations, Department of Economic and Social Affairs,
Population Division} "World Population Prospects: The 2015 Revision"
(\href{https://esa.un.org/unpd/wpp/}{website}).
}
\seealso{
\code{\link[madrat:readSource]{madrat::readSource()}}

\code{\link[=downloadUN_PopDiv]{downloadUN_PopDiv()}}
}
