% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readWDI.R
\name{downloadWDI}
\alias{downloadWDI}
\alias{readWDI}
\alias{convertWDI}
\title{Read WDI data}
\usage{
downloadWDI()

readWDI(subtype)

convertWDI(x, subtype)
}
\arguments{
\item{subtype}{A string. Type of WDI data that should be read. Use the
World Bank indicator abbreviation. Available subtypes are:
\itemize{
\item \code{"SP.POP.TOTL"}: Population, total
\item \code{"SP.POP.1564.TO"}: Working age population (15-64 years old)
\item \code{"SP.URB.TOTL.IN.ZS"}: Urban Population (\% of total)
\item \code{"NY.GDP.MKTP.PP.KD"}: GDP,PPP (constant 2017 international Dollar)
\item \code{"NV.AGR.TOTL.KD"}: Ag GDP, MER, (2010 US$)
\item \code{"PA.NUS.PPPC.RF"}: Price level ratio of PPP conversion factor (GDP) to market exchange rate
\item \code{"AG.SRF.TOTL.K2"}: Surface area (in square kms)
}}

\item{x}{MAgPIE object returned by readWDI}
}
\value{
The read-in data, usually a magpie object. If supplementary is TRUE a list including
the data and metadata is returned instead. The temporal and data dimensionality
should match the source data. The spatial dimension should either match the source data or,
if the convert argument is set to TRUE, should be on ISO code country level.
}
\description{
Download, read and convert WDI (World development indicators) data.
}
\examples{
\dontrun{
readSource("WDI", subtype = "SP.POP.TOTL")
}
}
\seealso{
\code{\link[madrat:readSource]{madrat::readSource()}} and \code{\link[madrat:downloadSource]{madrat::downloadSource()}}
}
