% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolCreateComparisonPlots.R
\name{toolCreateComparisonPlots}
\alias{toolCreateComparisonPlots}
\title{Plot differences between two versions}
\usage{
toolCreateComparisonPlots(new = "current", old = "7.2.1", tCutOff = 2050)
}
\arguments{
\item{new}{description}

\item{old}{description}

\item{tCutOff}{description}
}
\value{
An invisible list of plots
}
\description{
Create line plots comparing two different versions of the scenarios
}
\keyword{internal}
