% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolGetScenarioDefinition.R
\name{toolGetScenarioDefinition}
\alias{toolGetScenarioDefinition}
\title{Get information on available scenarios}
\usage{
toolGetScenarioDefinition(driver = NULL, scen = NULL, aslist = FALSE)
}
\arguments{
\item{driver}{NULL or a character vector designating the driver for which information is to be returned. If NULL,
information for all drivers is returned. Available drivers are:
\itemize{
\item GDP
\item Population
\item GDPpc
\item Labour
\item Urban
}}

\item{scen}{NULL or a character vector designating the scenario for which information is to be returned. If NULL,
information for all scenarios is returned.}

\item{aslist}{TRUE or FALSE (default). If TRUE then the pastData, futureData and harmonization strings are returned
as a list.}
}
\value{
A tibble with the driver and scenario information.
}
\description{
toolGetScenarioDefinition can be used to figure out which scenarios are made available by mrdrivers, and how they
are constructed, i.e. what past data, future data and harmonization methods are used.
}
\examples{
toolGetScenarioDefinition()
toolGetScenarioDefinition(driver = "GDP")
toolGetScenarioDefinition(scen = "SSP2EU")
toolGetScenarioDefinition(driver = "Population", scen = "SSPs", aslist = TRUE)

}
