% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolGetUnitDollar.R
\name{toolGetUnitDollar}
\alias{toolGetUnitDollar}
\title{Get monetary unit}
\usage{
toolGetUnitDollar(returnOnlyBase = FALSE, inPPP = FALSE)
}
\arguments{
\item{returnOnlyBase}{TRUE or FALSE (default). If true only the base year is returned (as string).}

\item{inPPP}{TRUE or FALSE (default). If TRUE the the string ends in 'Int$PPP', instead of 'Int$MER'.}
}
\value{
A string with the monetary unit: 'constant 2017 US$MER'.
}
\description{
toolGetUnitDollar returns unit used for monetary values.
}
\examples{
toolGetUnitDollar()
toolGetUnitDollar(inPPP = TRUE)
toolGetUnitDollar(returnOnlyBase = TRUE)

}
